+++ Content/paper.bib
@@ -9,6 +9,7 @@
 @article{Bisong_2019,
   title={Google colaboratory},
   author={Bisong, Ekaba and Bisong, Ekaba},
+  DOI={10.1007/978-1-4842-4470-8},
   journal={Building machine learning and deep learning models on google cloud platform: a comprehensive guide for beginners},
   pages={59--64},
   year={2019},
@@ -18,6 +19,7 @@
 @article{Boswell_2023,
   title={Smartphone videos of the sit-to-stand test predict osteoarthritis and health outcomes in a nationwide study},
   author={Boswell, Melissa A and Kidzi{\'n}ski, {\L}ukasz and Hicks, Jennifer L and Uhlrich, Scott D and Falisse, Antoine and Delp, Scott L},
+  DOI={10.1038/s41746-023-00775-1},
   journal={npj Digital Medicine},
   volume={6},
   number={1},
@@ -40,6 +42,7 @@
 @article{Bright_2012,
   title={Effect of clinical decision-support systems: a systematic review},
   author={Bright, Tiffani J and Wong, Anthony and Dhurjati, Ravi and Bristow, Erin and Bastian, Lori and Coeytaux, Remy R and Samsa, Gregory and Hasselblad, Vic and Williams, John W and Musty, Michael D and others},
+  DOI={10.7326/0003-4819-157-1-201207030-00450},
   journal={Annals of internal medicine},
   volume={157},
   number={1},
@@ -71,16 +74,6 @@
   publisher={IEEE}
 }
 
-@article{Butterworth_1930,
-  title={On the theory of filter amplifiers},
-  author={Butterworth, Stephen},
-  journal={Wireless Engineer},
-  volume={7},
-  number={6},
-  pages={536--541},
-  year={1930}
-}
-
 @article{Cleveland_1981,
   title={LOWESS: A program for smoothing scatterplots by robust locally weighted regression},
   author={Cleveland, William S},
@@ -107,6 +100,7 @@
 @article{Di_2021,
   title={A 2D qualitative movement assessment of a deceleration task detects football players with high knee joint loading},
   author={Di Paolo, Stefano and Zaffagnini, Stefano and Tosarelli, Filippo and Aggio, Fabrizio and Bragonzoni, Laura and Grassi, Alberto and Della Villa, Francesco},
+  DOI={10.1007/s00167-021-06709-2},
   journal={Knee Surgery, Sports Traumatology, Arthroscopy},
   volume={29},
   pages={4032--4040},
@@ -117,6 +111,7 @@
 @article{Kidzinski_2020,
   title={Deep neural networks enable quantitative movement analysis using single-camera videos},
   author={Kidzi{\'n}ski, {\L}ukasz and Yang, Bryan and Hicks, Jennifer L and Rajagopal, Apoorva and Delp, Scott L and Schwartz, Michael H},
+  DOI={10.1038/s41467-020-17807-z},
   journal={Nature communications},
   volume={11},
   number={1},
@@ -150,6 +145,7 @@
 @article{Minssen_2020,
   title={Clinical trial data transparency and GDPR compliance: Implications for data sharing and open innovation},
   author={Minssen, Timo and Rajam, Neethu and Bogers, Marcel},
+  DOI={10.2139/ssrn.3413035},
   journal={Science and Public Policy},
   volume={47},
   number={5},
@@ -195,6 +191,7 @@
 @article{Patrizi_2016,
   title={Comparison between low-cost marker-less and high-end marker-based motion capture systems for the computer-aided assessment of working ergonomics},
   author={Patrizi, Alfredo and Pennestr{\`\i}, Ettore and Valentini, Pier Paolo},
+  DOI={10.1080/00140139.2015.1057238},
[…]
