+++ Content/paper.bib
@@ -239,6 +239,18 @@
   publisher={Butterworth-Heinemann}
 }
 
+@article{Yang_2007,
+  title={Hypergravity resistance exercise: the use of artificial gravity as potential countermeasure to microgravity},
+  author={Yang, Yifan and Baker, Michael and Graf, Scott and Larson, Jennifer and Caiozzo, Vincent J},
+  DOI={10.1152/japplphysiol.00772.2007},
+  journal={Journal of applied physiology},
+  volume={103},
+  number={5},
+  pages={1879--1887},
+  year={2007},
+  publisher={American Physiological Society}
+}
+
 @article{Zheng_2023,
   title={Deep learning-based human pose estimation: A survey},
   author={Zheng, Ce and Wu, Wenhan and Chen, Chen and Yang, Taojiannan and Zhu, Sijie and Shen, Ju and Kehtarnavaz, Nasser and Shah, Mubarak},
