### 2024-02-20 #6375 @davidpagnon
From the above similar articles, it seems like LukasAdamowicz (Sit2StandPy) and SwathiSheshadri (pose-3d) could be a good fit for reviewing sports2d!
Maybe tuliofalmeida, melund, and/or alcantarar as well.
