### 2024-03-07 970df4e
solved wrapping angles and video rotation metadata
