+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = sports2d
-version = 0.3.0
+version = 0.3.3
 author = David Pagnon
 author_email = contact@david-pagnon.com
 description = Detect pose and compute 2D joint angles from a video.
@@ -29,7 +29,7 @@ project_urls =
     
 [options]
 zip_safe = False
-python_requires = <3.12
+python_requires =< 3.12
 install_requires =
     ipython
     toml
@@ -38,7 +38,7 @@ install_requires =
     scipy
     anytree
     opencv-python
-    mediapipe
+    mediapipe==1.10.0
     matplotlib
     PyQt5
     statsmodels
