### 2024-03-08 9a868d8
mediapipe specific version for macOS
