+++ setup.cfg
@@ -38,7 +38,7 @@ install_requires =
     scipy
     anytree
     opencv-python
-    mediapipe==1.10.0
+    mediapipe==0.10.0
     matplotlib
     PyQt5
     statsmodels
