### 2024-03-24 d703933
mediapipe conflicting versions
mediapipe > 3.10 cannot be installed with python > 3.10 on macOs
mediapipe 3.10 cannot be installed with python 3.11 on windows
