+++ .github/workflows/continuous-integration.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: Build on Win-MacOS-Ubuntu with Python 3.8-3.121
+name: Build on Win-MacOS-Ubuntu with Python 3.8-3.10
 
 on:
   push:
@@ -23,7 +23,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.8", "3.9", "3.10", "3.11"]
+        python-version: ["3.8", "3.9", "3.10"]
         include:
           - os: ubuntu-latest
             path: ~/.cache/pip
