+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = sports2d
-version = 0.3.3
+version = 0.3.4
 author = David Pagnon
 author_email = contact@david-pagnon.com
 description = Detect pose and compute 2D joint angles from a video.
@@ -38,7 +38,7 @@ install_requires =
     scipy
     anytree
     opencv-python
-    mediapipe==0.10.0
+    mediapipe
     matplotlib
     PyQt5
     statsmodels
