### 2024-03-24 b73ea98
mediapipe version issue with certain config
