+++ Sports2D/Utilities/skeletons.py
@@ -198,7 +198,7 @@ BODY_135 = Node("CHip", id=None, children=[
 ])
 
 
-'''BLAZEPOSE (full-body with simplified hand and foot, from mediapipe)
+'''BLAZEPOSE (full-body with simplified hand and foot, from mediapipe) 
 https://google.github.io/mediapipe/solutions/pose'''
 BLAZEPOSE = Node("CHip", id=None, children=[
     Node("RHip", id=24, children=[
