### 2024-03-24 5d11e96
test build with specific mediapipe version
