+++ setup.cfg
@@ -38,7 +38,7 @@ install_requires =
     scipy
     anytree
     opencv-python
-    mediapipe
+    mediapipe==0.10.9
     matplotlib
     PyQt5
     statsmodels
