### 2024-03-24 #6375 @davidpagnon
Hi Kevin, thank you for offering yourself as an editor!

I suggested a few in a previous comment, but of course, I'm open to anyone else.

> From the above similar articles, it seems like LukasAdamowicz (Sit2StandPy) and SwathiSheshadri (pose-3d) could be a good fit for reviewing sports2d! Maybe tuliofalmeida, melund, and/or alcantarar as well.

P.S.: I just edited a github action, build was unsuccessful on certain distribution due to an incompatible package version. All builds are now successful.
