### 2024-05-27 #6375 @Kevin-Mattheus-Moerman
@davidpagnon thanks for suggesting the authors of this submission: https://github.com/openjournals/joss-reviews/issues/6627 as reviewers. Although under different circumstances they could be suitable reviewers, because you are reviewer their submission they cannot currently review yours as that would constitute a potential conflict of interest.

Hence @rmndrs89 and colleagues, unfortunately you cannot be a reviewer for this work. Thanks for offering to review though.
