### 2024-05-30 #6375 @Kevin-Mattheus-Moerman
@SwathiSheshadri, @CVHammond, @destogl, @moorepants, @modenaxe, @LukasAdamowicz @tuliofalmeida @alcantarar

would you be interested in helping to review this submission for JOSS? The work features 2D join angle analysis from sport video footage.

Software: https://github.com/davidpagnon/Sports2D
Paper: https://raw.githubusercontent.com/openjournals/joss-papers/joss.06375/joss.06375/10.21105.joss.06375.pdf
