### 2024-06-05 #6849 @editorialbot
[REVIEW]: Sports2D: Compute 2D joint and segment angles from your smartphone
**Submitting author:** @DavidPagnon (<a href="http://orcid.org/0000-0002-6891-8331">David Pagnon</a>)
**Repository:** https://github.com/davidpagnon/Sports2D
**Branch with paper.md** (empty if default branch):
**Version:** v0.4.6
