### 2024-07-03 davidpagnon/Sports2D#9 @tuliofalmeida
Review Joss - Tulio
Hey David,

First of all, I'm happy to see your work and [review it](https://github.com/openjournals/joss-reviews/issues/6849). I'm a physical therapist and I really think that Sports2D can be super useful for clinicians.
I was reading and using your colab notebook and everything worked super well with the example file.
I tried to use one old video that I have measured it using others software, but I had a small bug (maybe I did something wrong).
I used the upload button and I the data was fully uploaded but didn't appeared in the folder videos, on purpose I tried a file with spaces i the name ( gait patient 1.mp4 - as a clinician would do). Then I removed the spaces but didn't worked also:
![image](https://github.com/davidpagnon/Sports2D/assets/54067432/de1b071d-099f-4262-b732-c8c2f4ca984b)

I did something wrong? Also, it's possible to previously upload this data in Google Drive use the folder path in Sports2D ? Maybe can be nice to receive the path of a folder (maybe the person already have it).

Thank you,

Tulio
