### 2024-07-03 davidpagnon/Sports2D#9 @davidpagnon
Interesting, thanks!
I'll read it more carefully and fix it as soon as possible, but in the meantime you can write ['test.mp4'] instead and it should work.
