### 2024-07-10 ba231c6
handles other pose models better
