### 2024-07-15 #6849 @davidpagnon
Hi, thank you Nicos and the other reviewers.

To @Kevin-Mattheus-Moerman: @hunminkim98 and I (mostly him) are currently working to make it run in real time, I wonder if it's worth completing the review with the library the way it is, or if it's better to wait?

Note that it could be a couple months because it's not the hottest priority right now.

https://github.com/user-attachments/assets/ffc549b4-f76f-4fa5-a7d8-6bfae7e702ba
