+++ Sports2D/RTMPose.py
@@ -1,0 +1,692 @@
+import cv2
+import numpy as np
+from rtmlib import PoseTracker, BodyWithFeet, draw_skeleton
+import pandas as pd
+from Sports2D.compute_angles import (
+    joint_angles_series_from_points, 
+    segment_angles_series_from_points, 
+    overlay_angles, 
+    display_figures_fun
+)
+from Sports2D.detect_pose import (
+    draw_bounding_box,
+    draw_keypts_skel,
+    save_imgvid_reID,
+    json_to_csv
+)
+from Sports2D.Utilities import filter
+import math
+import os
+import glob
+import logging
+from tqdm import tqdm
+import json
+import re
+from datetime import datetime
+from pathlib import Path
+# Define joint angle parameters
+def get_joint_angle_params(joint):
+    joint_angle_dict = {
+        'Right ankle': [['right_heel', 'right_big_toe', 'right_ankle', 'right_knee'], 'dorsiflexion', -90, -1],
+        'Left ankle': [['left_heel', 'left_big_toe', 'left_ankle', 'left_knee'], 'dorsiflexion', -90, -1],
+        'Right knee': [['right_hip', 'right_knee', 'right_ankle'], 'flexion', -180, -1],
+        'Left knee': [['left_hip', 'left_knee', 'left_ankle'], 'flexion', -180, -1],
+        'Right hip': [['right_knee', 'right_hip', 'right_shoulder'], 'flexion', -180, -1],
+        'Left hip': [['left_knee', 'left_hip', 'left_shoulder'], 'flexion', -180, -1],
+        'Right shoulder': [['right_hip', 'right_shoulder', 'right_elbow'], 'flexion', 0, 1],
+        'Left shoulder': [['left_hip', 'left_shoulder', 'left_elbow'], 'flexion', 0, 1],
+        'Right elbow': [['right_wrist', 'right_elbow', 'right_shoulder'], 'flexion', -180, -1],
+        'Left elbow': [['left_wrist', 'left_elbow', 'left_shoulder'], 'flexion', -180, -1],
+    }
+    return joint_angle_dict.get(joint)
+
+# Define segment angle parameters
+def get_segment_angle_params(segment):
+    segment_angle_dict = {
+        'Right foot': [['right_heel', 'right_big_toe'], 'horizontal', 0, -1],
+        'Left foot': [['left_heel', 'left_big_toe'], 'horizontal', 0, -1],
+        'Right shank': [['right_knee', 'right_ankle'], 'horizontal', 0, -1],
+        'Left shank': [['left_knee', 'left_ankle'], 'horizontal', 0, -1],
+        'Right thigh': [['right_hip', 'right_knee'], 'horizontal', 0, -1],
+        'Left thigh': [['left_hip', 'left_knee'], 'horizontal', 0, -1],
+        'Trunk': [['right_shoulder', 'right_hip'], 'horizontal', 0, 1],
+        'Right arm': [['right_shoulder', 'right_elbow'], 'horizontal', 0, -1],
+        'Left arm': [['left_shoulder', 'left_elbow'], 'horizontal', 0, -1],
+        'Right forearm': [['right_elbow', 'right_wrist'], 'horizontal', 0, -1],
+        'Left forearm': [['left_elbow', 'left_wrist'], 'horizontal', 0, -1],
+    }
+    return segment_angle_dict.get(segment)
+
+# Save keypoints and scores to a JSON file in OpenPose format
+def save_to_openpose(json_file_path, keypoints, scores):
+    nb_detections = len(keypoints)
+    detections = []
+    for i in range(nb_detections):
+        keypoints_with_confidence_i = []
+        for kp, score in zip(keypoints[i], scores[i]):
[…]
