### 2024-07-16 #6849 @Kevin-Mattheus-Moerman
@davidpagnon I suggest we continue with the current state of things as submitted. If you've implemented the features before this review is over we can revisit this. In any case I recommend you implement it with a clear example for instance, so that testing this new functionality is quick and simple, and would not take a lot of additional time.
