+++ setup.cfg
@@ -42,6 +42,8 @@ install_requires =
     matplotlib
     PyQt5
     statsmodels
+    rtmlib
+    openvino
 packages = find:
 
 [options.package_data]

[…]
