+++ Sports2D/compute_angles.py
@@ -15,11 +15,16 @@
     Optionally saves images and video with overlaid angles.
 
     Joint angle conventions:
-    - Ankle dorsiflexion: Between heel and big toe, and ankle and knee
-    - Knee flexion: Between hip, knee, and ankle 
-    - Hip flexion: Between knee, hip, and shoulder
-    - Shoulder flexion: Between hip, shoulder, and elbow
-    - Elbow flexion: Between wrist, elbow, and shoulder
+    - Ankle dorsiflexion: Between heel and big toe, and ankle and knee.\
+    *-90° when the foot is aligned with the shank.*
+    - Knee flexion: Between hip, knee, and ankle.\
+    *0° when the shank is aligned with the thigh.*
+    - Hip flexion: Between knee, hip, and shoulder.\
+    *0° when the trunk is aligned with the thigh.* 
+    - Shoulder flexion: Between hip, shoulder, and elbow.\
+    *180° when the arm is aligned with the trunk.*
+    - Elbow flexion: Between wrist, elbow, and shoulder.\
+    *0° when the forearm is aligned with the arm.*
 
     Segment angle conventions:
     Angles are measured anticlockwise between the horizontal and the segment.
@@ -342,11 +347,16 @@ def compute_angles_fun(config_dict, video_file):
     Optionally displays figures.
 
     Joint angle conventions:
-    - Ankle dorsiflexion: Between heel and big toe, and ankle and knee
-    - Knee flexion: Between hip, knee, and ankle 
-    - Hip flexion: Between knee, hip, and shoulder
-    - Shoulder flexion: Between hip, shoulder, and elbow
-    - Elbow flexion: Between wrist, elbow, and shoulder
+    - Ankle dorsiflexion: Between heel and big toe, and ankle and knee.\
+    *-90° when the foot is aligned with the shank.*
+    - Knee flexion: Between hip, knee, and ankle.\
+    *0° when the shank is aligned with the thigh.*
+    - Hip flexion: Between knee, hip, and shoulder.\
+    *0° when the trunk is aligned with the thigh.* 
+    - Shoulder flexion: Between hip, shoulder, and elbow.\
+    *180° when the arm is aligned with the trunk.*
+    - Elbow flexion: Between wrist, elbow, and shoulder.\
+    *0° when the forearm is aligned with the arm.*
 
     Segment angle conventions:
     Angles are measured anticlockwise between the horizontal and the segment.
