+++ build/lib/Sports2D/compute_angles.py
@@ -1,529 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-
-'''
-    ##############################################################
-    ## Compute joint and segment angles from csv position files ##
-    ##############################################################
-
-    Compute joint and segment angles from csv position files.
-    Automatically adjust angles when person switches to face the other way.
-    Save a 2D csv angle file per person.
-    Optionally filters results with Butterworth, gaussian, median, or loess filter.
-    Optionally displays figures.
-    Optionally saves images and video with overlaid angles.
-
-    Joint angle conventions:
-    - Ankle dorsiflexion: Between heel and big toe, and ankle and knee
-    - Knee flexion: Between hip, knee, and ankle 
-    - Hip flexion: Between knee, hip, and shoulder
-    - Shoulder flexion: Between hip, shoulder, and elbow
-    - Elbow flexion: Between wrist, elbow, and shoulder
-
-    Segment angle conventions:
-    Angles are measured anticlockwise between the horizontal and the segment.
-    - Foot: Between heel and big toe
-    - Shank: Between ankle and knee
-    - Thigh: Between hip and knee
-    - Arm: Between shoulder and elbow
-    - Forearm: Between elbow and wrist
-    - Trunk: Between hip midpoint and shoulder midpoint
-    
-    /!\ Warning /!\
-    - The angle estimation is only as good as the pose estimation algorithm, i.e., it is not perfect.
-    - It will lead to reliable results only if the persons move in the 2D plane (sagittal plane).
-    - The persons need to be filmed as perpendicularly as possible from their side.
-    If you need research-grade markerless joint kinematics, consider using several cameras,
-    and constraining angles to a biomechanically accurate model. See Pose2Sim for example: 
-    https://github.com/perfanalytics/pose2sim
-    
-    INPUTS:
-    - one or several position csv files
-    - a dictionary obtained from a configuration file (.toml extension)
-    - a skeleton model
-    
-    OUTPUTS:
-    - one csv file for joint and segment angles per detected person
-    - a logs.txt file 
-
-'''    
-
-
-## INIT
-import logging
-import os
-from pathlib import Path
-import pandas as pd
-import numpy as np
-import cv2
-import matplotlib.pyplot as plt
-from Sports2D.Sports2D import base_params
-from Sports2D.Utilities import filter, common
-from Sports2D.Utilities.skeletons import *
-
-
-## AUTHORSHIP INFORMATION
-__author__ = "David Pagnon"
-__copyright__ = "Copyright 2023, Sports2D"
-__credits__ = ["David Pagnon"]
-__license__ = "BSD 3-Clause License"
-__version__ = "0.3.0"
-__maintainer__ = "David Pagnon"
-__email__ = "contact@david-pagnon.com"
-__status__ = "Development"
-
-
-
-# CONSTANTS
-# dict: name: points, extra, offset, invert. 
-# Most angles are multiplied by -1 because the OpenCV y axis points down.
-joint_angle_dict = { 
-    'Right ankle': [['RHeel', 'RBigToe', 'RAnkle', 'RKnee'], 'dorsiflexion', -90, -1],
-    'Left ankle': [['LHeel', 'LBigToe', 'LAnkle', 'LKnee'], 'dorsiflexion', -90, -1],
-    'Right knee': [['RHip', 'RKnee', 'RAnkle'], 'flexion', -180, -1],
-    'Left knee': [['LHip', 'LKnee', 'LAnkle'], 'flexion', -180, -1],
[…]
