+++ build/lib/Sports2D/Utilities/__init__.py
@@ -1,7 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-import sys
-
-__version__ = '0.1.0'
-VERSION = __version__
