+++ Sports2D/compute_angles.py
@@ -227,38 +227,19 @@ def flip_left_right_direction_webcam(df_points):
         if col not in df_points.columns:
             raise ValueError(f"Required column {col} is missing from the dataframe")
     
-    # Left foot 관련 정보 출력
-    left_foot_keypoints = ['left_big_toe', 'left_small_toe', 'left_heel', 'left_ankle']
-    print("\nLeft Foot Keypoint Information:")
-    for kp in left_foot_keypoints:
-        x = df_points[f'{kp}_x'].values[0]
-        y = df_points[f'{kp}_y'].values[0]
-        score = df_points[f'{kp}_score'].values[0] if f'{kp}_score' in df_points.columns else 'N/A'
-        print(f"{kp}: x={x:.2f}, y={y:.2f}, score={score}")
+    # Calculate orientation
+    right_orientation = df_points['right_big_toe_x'] - df_points['right_heel_x']
+    left_orientation = df_points['left_big_toe_x'] - df_points['left_heel_x']
+    orientation = right_orientation + left_orientation
     
-    right_direction = df_points['right_big_toe_x'] - df_points['right_heel_x']
-    left_direction = df_points['left_big_toe_x'] - df_points['left_heel_x']
+    # Flip X coordinates based on orientation
+    x_columns = [col for col in df_points.columns if col.endswith('_x')]
+    df_points[x_columns] = df_points[x_columns] * np.where(orientation >= 0, 1, -1).reshape(-1, 1)
     
-    print(f"\nRight foot direction: {right_direction.values[0]:.2f}")
-    print(f"Left foot direction: {left_direction.values[0]:.2f}")
-    
-    # 방향 결정 로직
-    if right_direction.values[0] < 0:
-        overall_direction = -1
-    elif left_direction.values[0] > 0:
-        overall_direction = 1
+    if (orientation < 0).any():
+        print("Some coordinates flipped")
     else:
-        overall_direction = 0
-    
-    print(f"Overall direction: {overall_direction}")
-    
-    # 좌표 반전 로직
-    if overall_direction < 0:
-        x_columns = [col for col in df_points.columns if col.endswith('_x')]
-        df_points[x_columns] = -df_points[x_columns]
-        print("Coordinates flipped")
-    else:
-        print("Coordinates not flipped")
+        print("No coordinates flipped")
     
     return df_points
 
@@ -279,16 +260,25 @@ def joint_angles_series_from_points(df_points, angle_params, kpt_thr):
     # print(f"Keypoints to use: {angle_params[0]}")
     
     keypt_series = []
-    print(f"df_points_c : {df_points.columns}")
+    # print(f"df_points columns: {df_points.columns}")
     
     for k in angle_params[0]:
-        if f"{k}_x" in df_points.columns and f"{k}_y" in df_points.columns:
-            keypt = df_points[[f"{k}_x", f"{k}_y"]]
-            # print(f"Keypoint {k} added: {keypt.values[0]}")
-            keypt_series.append(keypt)
+        if f"{k}_x" in df_points.columns and f"{k}_y" in df_points.columns and f"{k}_score" in df_points.columns:
+            score = df_points[f"{k}_score"].values[0]
+            if score >= kpt_thr:
+                keypt = df_points[[f"{k}_x", f"{k}_y"]]
+                # print(f"Keypoint {k} added: {keypt.values[0]}, score: {score}")
+                keypt_series.append(keypt)
+            else:
[…]
