+++ Sports2D/Demo/Config_demo.toml
@@ -22,8 +22,8 @@ result_dir = '' # BETWEEN SINGLE QUOTES! # If empty, project dir is current dir
 data_type = "webcam"
 video_path = "path/to/your/video.mp4"
 det_frequency = 1
-mode = "performance"
-keypoints_threshold = 0.3
+mode = "balanced"
+keypoints_threshold = 0.4
 tracking = false
 to_openpose = false
 display_detection = true
