+++ Sports2D/compute_angles.py
@@ -62,7 +62,7 @@ import matplotlib.pyplot as plt
 import traceback
 from Sports2D.Sports2D import base_params
 from Sports2D.Utilities import filter, common
-from Sports2D.Utilities.skeletons import *
+from Sports2D.Utilities.skeletons import halpe26_rtm
 
 
 
@@ -195,30 +195,16 @@ def flip_left_right_direction(df_points):
     OUTPUT:
     - df_points: dataframe of pose detection with flipped X coordinates
     '''
-    
-    # Check if the required columns exist
-    required_columns = ['right_big_toe', 'right_heel', 'left_big_toe', 'left_heel']
-    for col in required_columns:
-        if col not in df_points.columns.get_level_values(2):
-            print(f"Debug: Column '{col}' not found in df_points")
-    
-    try:
-        righ_orientation = df_points.iloc[:,df_points.columns.get_level_values(2)=='right_big_toe'].iloc[:,0] - df_points.iloc[:,df_points.columns.get_level_values(2)=='right_heel'].iloc[:,0]
-    except Exception as e:
-        righ_orientation = pd.Series([0] * len(df_points))
-    
-    try:
-        left_orientation = df_points.iloc[:,df_points.columns.get_level_values(2)=='left_big_toe'].iloc[:,0] - df_points.iloc[:,df_points.columns.get_level_values(2)=='left_heel'].iloc[:,0]
-    except Exception as e:
-        left_orientation = pd.Series([0] * len(df_points))
-    
+    # print(f"dataframe columns: {df_points.columns}")
+    righ_orientation = df_points.iloc[:,df_points.columns.get_level_values(2)=='right_big_toe'].iloc[:,0] - df_points.iloc[:,df_points.columns.get_level_values(2)=='right_heel'].iloc[:,0]
+    # print(f"righ_orientation: {righ_orientation}")
+    left_orientation = df_points.iloc[:,df_points.columns.get_level_values(2)=='left_big_toe'].iloc[:,0] - df_points.iloc[:,df_points.columns.get_level_values(2)=='left_heel'].iloc[:,0]
+    # print(f"left_orientation: {left_orientation}")
     orientation = righ_orientation + left_orientation
+    # print(f"orientation: {orientation}")
+    df_points.iloc[:,2::3] = df_points.iloc[:,2::3] * np.where(orientation>=0, 1, -1).reshape(-1,1)
 
-    try:
-        df_points.iloc[:,2::3] = df_points.iloc[:,2::3] * np.where(orientation>=0, 1, -1).reshape(-1,1)
-    except Exception as e:
-        print(f"Debug: Error in flipping coordinates: {str(e)}")
-
+    # print("successfully flipped left/right direction")
     return df_points
 
 def flip_left_right_direction_webcam(df_points):
@@ -256,9 +242,6 @@ def joint_angles_series_from_points(df_points, angle_params, kpt_thr):
     '''
     
     # Retrieve points
-    # print(f"Processing joint: {angle_params[1]}")
-    # print(f"Keypoints to use: {angle_params[0]}")
-    
     keypt_series = []
     # print(f"df_points columns: {df_points.columns}")
     
@@ -393,11 +376,6 @@ def joint_angles_series_from_csv(df_points, angle_params, kpt_thr):
     
     if ang_series.mean() > 180: ang_series -= 360
     if ang_series.mean() < -180: ang_series += 360
-    
-    ang_series = np.abs(ang_series)
-    if ang_series is None or len(ang_series) == 0:
-        return None
-    return ang_series
 
     return ang_series
 
@@ -420,7 +398,6 @@ def segment_angles_series_from_csv(df_points, angle_params, segment, kpt_thr):
         try:
             score_series = df_points.xs((k, 'score'), level=[2, 3], axis=1)
[…]
