+++ Sports2D/compute_angles.py
@@ -81,7 +81,17 @@ __status__ = "Development"
 # CONSTANTS
 # dict: name: points, extra, offset, invert. 
 # Most angles are multiplied by -1 because the OpenCV y axis points down.
+# define joint angle parameters
 def get_joint_angle_params(joint):
+    """
+    Returns parameters for joint angle calculations.
+    
+    Args:
+    joint (str): Name of the joint
+
+    Returns:
+    list: Parameters for angle calculation [keypoints, type, offset, invert]
+    """
     joint_angle_dict = {
         'Right ankle': [['right_heel', 'right_big_toe', 'right_ankle', 'right_knee'], 'dorsiflexion', -90, -1],
         'Left ankle': [['left_heel', 'left_big_toe', 'left_ankle', 'left_knee'], 'dorsiflexion', -90, -1],
@@ -98,6 +108,15 @@ def get_joint_angle_params(joint):
 
 # Define segment angle parameters
 def get_segment_angle_params(segment):
+    """
+    Returns parameters for segment angle calculations.
+    
+    Args:
+    segment (str): Name of the segment
+
+    Returns:
+    list: Parameters for angle calculation [keypoints, type, offset, invert]
+    """
     segment_angle_dict = {
         'Right foot': [['right_heel', 'right_big_toe'], 'horizontal', 0, -1],
         'Left foot': [['left_heel', 'left_big_toe'], 'horizontal', 0, -1],
@@ -114,7 +133,7 @@ def get_segment_angle_params(segment):
     return segment_angle_dict.get(segment)
 
 # FUNCTIONS
-def display_figures_fun(df_list):
+def display_figures_fun_ang(df_list):
     '''
     Displays filtered and unfiltered data for comparison
     /!\ Crashes on the third window...
@@ -156,10 +175,7 @@ def points2D_to_angles(points_list):
     If parameters are float, returns a float between 0.0 and 360.0
     If parameters are arrays, returns an array of floats between 0.0 and 360.0
     '''
-    # print(f"points2D_to_angles received points_list: {points_list}")
-    # print(f"Number of points: {len(points_list)}")
-
-    if len(points_list) < 2:
+    if len(points_list) < 2: # if not enough points, return None
         return None
     
     ax, ay = points_list[0]
@@ -184,49 +200,47 @@ def points2D_to_angles(points_list):
 
     return ang_deg
 
-def flip_left_right_direction(df_points):
+def flip_left_right_direction(df_points, data_type):
     '''
     Inverts X coordinates to get consistent angles when person changes direction and goes to the left.
     The person is deemed to go to the left when their toes are to the left of their heels.
     
     INPUT:
     - df_points: dataframe of pose detection
+    - data_type: 'video' or 'webcam'
     
     OUTPUT:
     - df_points: dataframe of pose detection with flipped X coordinates
     '''
-    # print(f"dataframe columns: {df_points.columns}")
-    righ_orientation = df_points.iloc[:,df_points.columns.get_level_values(2)=='right_big_toe'].iloc[:,0] - df_points.iloc[:,df_points.columns.get_level_values(2)=='right_heel'].iloc[:,0]
-    # print(f"righ_orientation: {righ_orientation}")
-    left_orientation = df_points.iloc[:,df_points.columns.get_level_values(2)=='left_big_toe'].iloc[:,0] - df_points.iloc[:,df_points.columns.get_level_values(2)=='left_heel'].iloc[:,0]
-    # print(f"left_orientation: {left_orientation}")
-    orientation = righ_orientation + left_orientation
-    # print(f"orientation: {orientation}")
[…]
