+++ Sports2D/detect_pose.py
@@ -62,10 +62,13 @@ from Sports2D.compute_angles import (
     get_joint_angle_params,
     get_segment_angle_params, 
     joint_angles_series_from_points, 
-    segment_angles_series_from_points, 
+    segment_angles_series_from_points,
+    joint_angles_series_from_csv,
+    segment_angles_series_from_csv, 
     draw_bounding_box,
     overlay_angles,
-    flip_left_right_direction_webcam)
+    flip_left_right_direction,
+    display_figures_fun_ang)
 from Sports2D.Utilities import filter, common
 from Sports2D.Utilities.skeletons import halpe26_rtm
 from rtmlib import PoseTracker, BodyWithFeet, draw_skeleton
@@ -84,7 +87,7 @@ __status__ = "Development"
 
 
 # FUNCTIONS
-def display_figures_fun(df_list):
+def display_figures_fun_cords(df_list):
     '''
     Displays filtered and unfiltered data for comparison
     /!\ Crashes on the third window...
@@ -249,17 +252,18 @@ def save_to_openpose(json_file_path, keypoints, scores):
         json.dump(json_output, json_file)
 
 
-def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, show_plots):
+def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, show_plots, min_detection_time=1):
     '''
     Converts frame-by-frame json coordinate files 
     to one csv files per detected person
 
     INPUTS:
     - json_path: directory path of json files
-    - pose_model: string, to get tree from skeletons.py
+    - frame_rate: frame rate of the video
     - interp_gap_smaller_than: integer, maximum number of missing frames for conducting interpolation
     - filter_options: list, options for filtering
     - show_plots: boolean, show plots or not
+    - min_detection_time: minimum detection time in seconds (default: 1 second)
 
     OUTPUTS:
     - Creation of one csv files per detected person
@@ -275,7 +279,6 @@ def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, 
     # Retrieve coordinates
     logging.info('Sorting people across frames.')
     json_fnames = sorted(list(json_path.glob('*.json')))
-    # print(f"nb json files: {len(json_fnames)}")
     nb_persons_to_detect = max([len(json.load(open(json_fname))['people']) for json_fname in json_fnames])
     Coords = [np.array([]).reshape(0,keypoints_nb*3)] * nb_persons_to_detect
     for json_fname in json_fnames:    # for each frame
@@ -303,6 +306,15 @@ def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, 
     
     # Inject coordinates in dataframes and save
     for i in range(nb_persons_to_detect): 
+        # Calculate detection time
+        detection_mask = np.any(Coords[i] != 0, axis=1)
+        detection_time = np.sum(detection_mask) / frame_rate
+
+        # Skip if detected for less than min_detection_time
+        if detection_time < min_detection_time:
+            logging.info(f'Person {i}: Detected for less than {min_detection_time} second(s). Skipping.')
+            continue
+
         # Prepare csv header
         scorer = ['DavidPagnon']*(keypoints_nb*3+1)
         individuals = [f'person{i}']*(keypoints_nb*3+1)
@@ -347,38 +359,9 @@ def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, 
         # Display figures
         if show_plots:
             logging.info(f'Person {i}: Displaying figures.')
-            display_figures_fun(df_list)
-            
-
-
-
-def rewrite_vid(video_path):
-    '''
-    Rewrite video
[…]
