### 2024-07-22 8d0a6e5
fix bug of person_id in bounding box
