+++ Sports2D/detect_pose.py
@@ -408,7 +408,7 @@ def process_video(video_path, video_result_path,pose_tracker, tracking, output_f
     except:
         raise NameError(f"{video_path} is not a video. Images must be put in one subdirectory per camera.")
     
-    pose_dir = os.path.abspath(os.path.join(video_result_path, '..', 'pose'))
+    pose_dir = os.path.abspath(os.path.join(video_result_path, '..', 'video_results'))
     print(f"pose_dir: {pose_dir}")
     if not os.path.isdir(pose_dir): os.makedirs(pose_dir)
     video_name_wo_ext = os.path.splitext(os.path.basename(video_path))[0]
@@ -487,7 +487,7 @@ def process_video(video_path, video_result_path,pose_tracker, tracking, output_f
 
 # If input is a webcam
 def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segment_angles, save_video, save_images, interp_gap_smaller_than, 
-                   filter_options, show_plots, flip_left_right, kpt_thr, data_type):
+                   filter_options, show_plots, flip_left_right, kpt_thr, data_type, do_filter_angles, min_detection_time):
     """
     Process a live webcam feed to detect poses and calculate angles.
 
@@ -550,7 +550,7 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
 
     # Output directory setup
     current_time = datetime.now().strftime("%Y%m%d_%H%M%S")
-    output_dir = Path(os.path.join(os.getcwd(), f'realtime_pose_output_{current_time}'))
+    output_dir = Path(os.path.join(os.getcwd(), f'webcam_results_{current_time}'))
     output_dir.mkdir(parents=True, exist_ok=True)
     json_output_dir = output_dir / 'json'
     json_output_dir.mkdir(parents=True, exist_ok=True)
@@ -558,7 +558,7 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
     # Video writer setup
     video_writer = None
     if save_video:
-        video_output_path = str(output_dir / 'output_video.mp4')
+        video_output_path = str(output_dir / 'webcam_video.mp4')
         fourcc = cv2.VideoWriter_fourcc(*'mp4v')
         video_writer = cv2.VideoWriter(video_output_path, fourcc, frame_rate, (width, height))
         if not video_writer.isOpened():
@@ -567,7 +567,7 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
 
     # Image output directory setup
     if save_images:
-        img_output_dir = output_dir / 'images'
+        img_output_dir = output_dir / 'webcam_images'
         img_output_dir.mkdir(parents=True, exist_ok=True)
     
     frame_count = 0
@@ -684,7 +684,7 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
         filter_options[4] = frame_rate
         filter_options = tuple(filter_options)
     
-    json_to_csv(json_output_dir, frame_rate, interp_gap_smaller_than, filter_options, show_plots, min_detection_time=1)
+    json_to_csv(json_output_dir, frame_rate, interp_gap_smaller_than, filter_options, show_plots, min_detection_time)
 
     # Recalculate angles from filtered CSV files and apply filtering
     for person_idx in keypoints_data.keys():
@@ -755,7 +755,7 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
             df_angles = [pd.DataFrame(np.array(angle_series).T, columns=index_angs_csv)]
 
             # Apply filtering to angles
-            if filter_options[0]:
+            if do_filter_angles:
                 filter_type = filter_options[1]
[…]
