+++ Sports2D/detect_pose.py
@@ -421,10 +421,6 @@ def process_video(video_path, video_result_path,pose_tracker, tracking, output_f
         fps = cap.get(cv2.CAP_PROP_FPS) # Get the frame rate from the raw video
         W, H = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH)),int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT)) # Get the width and height from the raw video
         out = cv2.VideoWriter(output_video_path, fourcc, fps, (W, H)) # Create the output video file
-        
-    if display_detection:
-        if 'google.colab' not in sys.modules:
-            cv2.namedWindow(f"Pose Estimation {os.path.basename(video_path)}", cv2.WINDOW_NORMAL + cv2.WINDOW_KEEPRATIO)
 
     frame_idx = 0
     cap = cv2.VideoCapture(video_path)
@@ -531,22 +527,14 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
     Note: The function can be interrupted by pressing 'q' or using a keyboard interrupt (Ctrl+C).
     """
 
-    # Webcam setup
-    cam_id, cam_width, cam_height = webcam_settings
-    cap = cv2.VideoCapture(cam_id)
-    if not cap.isOpened():
-        print("Error: Could not open webcam.")
-        return
-    cap.set(cv2.CAP_PROP_FRAME_WIDTH, cam_width)
-    cap.set(cv2.CAP_PROP_FRAME_HEIGHT, cam_height)
-    
-    width = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH))
-    height = int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
-    print(f"Resolution: {width}x{height}")
-
-    cv2.namedWindow("Real-time Analysis", cv2.WINDOW_NORMAL)
-    frame_rate = cap.get(cv2.CAP_PROP_FPS) or 30
-    print(f"Webcam frame rate: {frame_rate}")
+    is_colab = 'google.colab' in sys.modules
+    if is_colab:
+        from IPython.display import display, Javascript, clear_output
+        from google.colab.output import eval_js
+        from base64 import b64decode
+        import matplotlib.pyplot as plt
+        import PIL.Image
+        import io
 
     # Output directory setup
     current_time = datetime.now().strftime("%Y%m%d_%H%M%S")
@@ -555,12 +543,88 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
     json_output_dir = output_dir / 'json'
     json_output_dir.mkdir(parents=True, exist_ok=True)
 
+    if is_colab:
+        cam_id, cam_width, cam_height = webcam_settings
+        frame_rate = 30  # Hardcoded fps
+
+        # Colab webcam setup
+        js = Javascript('''
+        async function setupWebcam() {
+          const video = document.createElement('video');
+          video.style.display = 'none';
+          const stream = await navigator.mediaDevices.getUserMedia({video: true});
+          video.srcObject = stream;
+          await video.play();
+          
+          const canvas = document.createElement('canvas');
+          canvas.width = video.videoWidth;
+          canvas.height = video.videoHeight;
+          const ctx = canvas.getContext('2d');
+          
+          return {video, canvas, ctx, stream, width: video.videoWidth, height: video.videoHeight};
+        }
+        
+        async function captureFrame(video, canvas, ctx) {
+          ctx.drawImage(video, 0, 0);
+          return canvas.toDataURL('image/jpeg');
+        }
+        
+        var webCamSetup = null;
+        
+        window.startWebcam = async function() {
+          if (!webCamSetup) {
[…]
