### 2024-07-22 #6849 @nicos1993
Hello @Kevin-Mattheus-Moerman,

I have tested the software using installation option 2 and followed the example, I can confirm the instructions are very clear and the demonstration ran smoothly with no issues. I believe this software will be very useful for coaches or practitioners wishing to perform a 2D automated kinematic analysis.

I do have a couple of suggestions that the author may wish to consider.
1. The demo video in the online repository and Figure 1 in the accompanying paper. Whilst the author acknowledges the limitation of using the software to analyze non-sagittal/frontal videos, the demo video/figure is very much non-planar and I would recommend using a demo video/figure which is specifically from either the sagittal plane or the frontal plane. In the software examples perhaps it would be nice to have both examples.
2. Segment and joint angle conventions. The author provides a citation for the convention of the angles used, but I would recommend using a more common and intuitive convention (e.g., from Winter 2009, Biomechanics and motor control of human movement). I believe this would also be more beneficial to coaches/practitioners. Lastly, in the figure I would consider including the definitions of the segment angles together with the joint angles.

Best wishes,

Nicos Haralabidis
