+++ Sports2D/compute_angles.py
@@ -15,16 +15,11 @@
     Optionally saves images and video with overlaid angles.
 
     Joint angle conventions:
-    - Ankle dorsiflexion: Between heel and big toe, and ankle and knee.\
-    *-90° when the foot is aligned with the shank.*
-    - Knee flexion: Between hip, knee, and ankle.\
-    *0° when the shank is aligned with the thigh.*
-    - Hip flexion: Between knee, hip, and shoulder.\
-    *0° when the trunk is aligned with the thigh.* 
-    - Shoulder flexion: Between hip, shoulder, and elbow.\
-    *180° when the arm is aligned with the trunk.*
-    - Elbow flexion: Between wrist, elbow, and shoulder.\
-    *0° when the forearm is aligned with the arm.*
+    - Ankle dorsiflexion: Between heel and big toe, and ankle and knee
+    - Knee flexion: Between hip, knee, and ankle 
+    - Hip flexion: Between knee, hip, and shoulder
+    - Shoulder flexion: Between hip, shoulder, and elbow
+    - Elbow flexion: Between wrist, elbow, and shoulder
 
     Segment angle conventions:
     Angles are measured anticlockwise between the horizontal and the segment.
@@ -514,14 +509,14 @@ def draw_joint_angle(frame, joint, angle, keypoints, scores, kpt_thr):
         It only draws the angle if all relevant keypoint scores are above the threshold.
     """
     joint_to_keypoints = {
-        "Right ankle": [14, 16, 21],  # knee, ankle, big_toe
-        "Left ankle": [13, 15, 20],   # knee, ankle, big_toe
-        "Right knee": [12, 14, 16],   # hip, knee, ankle
-        "Left knee": [11, 13, 15],    # hip, knee, ankle
-        "Right hip": [12, 12, 14],    # hip center, hip, knee
-        "Left hip": [11, 11, 13],     # hip center, hip, knee
-        "Right shoulder": [18, 6, 8], # neck, shoulder, elbow
-        "Left shoulder": [18, 5, 7],  # neck, shoulder, elbow
+        "Right ankle": [14, 16, 25, 21],  # right knee, right ankle, right heel, right big_toe
+        "Left ankle": [13, 15, 24, 20],   # left knee, left ankle, left heel, left big_toe
+        "Right knee": [12, 14, 16],   # right hip, right knee, right ankle
+        "Left knee": [11, 13, 15],    # left hip, knee, ankle
+        "Right hip": [18, 12, 14],    # neck, right hip, right knee
+        "Left hip": [18, 11, 13],     # neck, left hip, left knee
+        "Right shoulder": [19, 6, 8], # neck, shoulder, elbow
+        "Left shoulder": [19, 5, 7],  # neck, shoulder, elbow
         "Right elbow": [6, 8, 10],    # shoulder, elbow, wrist
         "Left elbow": [5, 7, 9],      # shoulder, elbow, wrist
     }
@@ -530,10 +525,19 @@ def draw_joint_angle(frame, joint, angle, keypoints, scores, kpt_thr):
         pts = [keypoints[i] for i in joint_to_keypoints[joint]]
         scores_pts = [scores[i] for i in joint_to_keypoints[joint]]
         if all(score >= kpt_thr for score in scores_pts):
-            pt1, pt2, pt3 = pts
-            draw_angle_arc(frame, joint, pt1, pt2, pt3, angle)
+            if 'ankle' in joint.lower():
+                draw_angle_arc(frame, joint, pts, angle)
+            else:
+                pt1, pt2, pt3 = pts
+                draw_angle_arc(frame, joint, [pt1, pt2, pt3], angle)
 
-def draw_angle_arc(frame, joint, pt1, pt2, pt3, angle):
[…]
