### 2024-07-23 #6849 @davidpagnon
Hi Nicos,

Thank you for your review!

1. I fully agree with your statement that the Demo video was probably not the best, given that it does not lie in the plane. I will replace it with a video of someone doing jumping jacks (frontal plane) and another person doing tuck jumps (sagittal plane).

2. I believe the joint angle convention is the same as the one from Winter 2009, but I agree that the way it is represented makes it confusing. I will add this figure to the article and the tutorial.
![joint_convention](https://github.com/user-attachments/assets/8d0a14dd-ca9e-4f3a-b406-d3fe78963c30)

I will ping you once the fixes have been made.
