### 2024-07-23 6a049e6
Change in figure and authors
