+++ Sports2D/detect_pose.py
@@ -45,19 +45,18 @@ import logging
 from pathlib import Path
 from sys import platform
 import json
-import subprocess
 import itertools as it
 import numpy as np
 import pandas as pd
 import cv2
 import matplotlib.pyplot as plt
 import matplotlib as mpl
-import re
 from tqdm import tqdm
 import torch
 import onnxruntime as ort
 from datetime import datetime
 import sys
+import time
 from IPython.display import clear_output
 
 from Sports2D.Sports2D import base_params
@@ -228,24 +227,35 @@ def sort_people(keyptpre, keypt, nb_persons_to_detect):
     
     return keypt
 
-def save_to_openpose(json_file_path, keypoints, scores):
+def save_to_openpose(json_file_path, keypoints, scores, kpt_thr):
+    score_threshold = kpt_thr
     nb_detections = len(keypoints)
+    print(f" number of detections: {nb_detections}")
     detections = []
+
     for i in range(nb_detections):
-        keypoints_with_confidence_i = []
-        for kp, score in zip(keypoints[i], scores[i]):
-            keypoints_with_confidence_i.extend([kp[0].item(), kp[1].item(), score.item()])
-        detections.append({
-            "person_id": [-1],
-            "pose_keypoints_2d": keypoints_with_confidence_i,
-            "face_keypoints_2d": [],
-            "hand_left_keypoints_2d": [],
-            "hand_right_keypoints_2d": [],
-            "pose_keypoints_3d": [],
-            "face_keypoints_3d": [],
-            "hand_left_keypoints_3d": [],
-            "hand_right_keypoints_3d": []
-        })
+        # 각 사람의 평균 score 계산
+        average_score = sum(score.item() for score in scores[i]) / len(scores[i])
+        
+        # 평균 score가 임계값 이상인 경우만 처리
+        if average_score >= score_threshold:
+            keypoints_with_confidence_i = []
+            for kp, score in zip(keypoints[i], scores[i]):
+                keypoints_with_confidence_i.extend([kp[0].item(), kp[1].item(), score.item()])
+            detections.append({
+                "person_id": [-1],
+                "pose_keypoints_2d": keypoints_with_confidence_i,
+                "face_keypoints_2d": [],
+                "hand_left_keypoints_2d": [],
+                "hand_right_keypoints_2d": [],
+                "pose_keypoints_3d": [],
+                "face_keypoints_3d": [],
+                "hand_left_keypoints_3d": [],
+                "hand_right_keypoints_3d": []
+            })
+            print(f"Person {i} passed the threshold. Average score: {average_score:.4f}")
+        # else:
+        #     print(f"Person {i} did not pass the threshold. Average score: {average_score:.4f}")
     
     json_output = {"version": 1.3, "people": detections}
     
@@ -279,11 +289,14 @@ def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, 
     keypoints_ids = [kp['id'] for kp in keypoints_info.values()]
     keypoints_names = [kp['name'] for kp in keypoints_info.values()]
     keypoints_nb = len(keypoints_ids)
+    print(f"keypoints_nb: {keypoints_nb}")
 
     # Retrieve coordinates
     logging.info('Sorting people across frames.')
     json_fnames = sorted(list(json_path.glob('*.json')))
+    # print(f"json_fnames: {json_fnames}")
     nb_persons_to_detect = max([len(json.load(open(json_fname))['people']) for json_fname in json_fnames])
+    print(f"nb_persons_to_detect: {nb_persons_to_detect}")
     Coords = [np.array([]).reshape(0,keypoints_nb*3)] * nb_persons_to_detect
[…]
