+++ Sports2D/compute_angles.py
@@ -124,7 +124,7 @@ def get_segment_angle_params(segment):
         'Left shank': [['left_knee', 'left_ankle'], 'horizontal', 0, -1],
         'Right thigh': [['right_hip', 'right_knee'], 'horizontal', 0, -1],
         'Left thigh': [['left_hip', 'left_knee'], 'horizontal', 0, -1],
-        'Trunk': [['right_shoulder', 'right_hip'], 'horizontal', 0, 1],
+        'Trunk': [['right_hip', 'right_shoulder'], 'horizontal', 0, -1],
         'Right arm': [['right_shoulder', 'right_elbow'], 'horizontal', 0, -1],
         'Left arm': [['left_shoulder', 'left_elbow'], 'horizontal', 0, -1],
         'Right forearm': [['right_elbow', 'right_wrist'], 'horizontal', 0, -1],
@@ -200,46 +200,92 @@ def points2D_to_angles(points_list):
 
     return ang_deg
 
-def flip_left_right_direction(df_points, data_type):
+def flip_left_right_direction(df_points, data_type):      # ! should check working on video and webcam (viedo ok, webcam not checked)
     '''
     Inverts X coordinates to get consistent angles when person changes direction and goes to the left.
     The person is deemed to go to the left when their toes are to the left of their heels.
     
     INPUT:
     - df_points: dataframe of pose detection
-    - data_type: 'video' or 'webcam'
     
     OUTPUT:
     - df_points: dataframe of pose detection with flipped X coordinates
     '''
-    
+
+    print(f"Flipping left-right direction for {data_type} data")
     if data_type not in ['video', 'webcam']:
         raise ValueError("data_type must be either 'video' or 'webcam'")
     
-    if data_type == 'video':
-        right_toe = df_points.iloc[:,df_points.columns.get_level_values(2)=='right_big_toe'].iloc[:,0]
-        right_heel = df_points.iloc[:,df_points.columns.get_level_values(2)=='right_heel'].iloc[:,0]
-        left_toe = df_points.iloc[:,df_points.columns.get_level_values(2)=='left_big_toe'].iloc[:,0]
-        left_heel = df_points.iloc[:,df_points.columns.get_level_values(2)=='left_heel'].iloc[:,0]
-    else:  # webcam
-        required_columns = ['right_big_toe_x', 'right_heel_x', 'left_big_toe_x', 'left_heel_x']
-        for col in required_columns:
-            if col not in df_points.columns:
-                raise ValueError(f"Required column {col} is missing from the dataframe")
-        right_toe = df_points['right_big_toe_x']
-        right_heel = df_points['right_heel_x']
-        left_toe = df_points['left_big_toe_x']
-        left_heel = df_points['left_heel_x']
+    # print('Flipping left-right direction...')
+    # print("DataFrame columns:", df_points.columns)
+    # print("Index levels:", df_points.columns.names)
     
-    right_orientation = right_toe - right_heel
-    left_orientation = left_toe - left_heel
-    orientation = right_orientation + left_orientation
+    # Set the level names explicitly if they are None
+    if df_points.columns.names == [None, None, None, None]:
+        df_points.columns.names = ['scorer', 'individuals', 'bodyparts', 'coords']
     
-    if data_type == 'video':
-        df_points.iloc[:,2::3] = df_points.iloc[:,2::3] * np.where(orientation>=0, 1, -1).reshape(-1,1)
-    else:  # webcam
-        x_columns = [col for col in df_points.columns if col.endswith('_x')]
-        df_points[x_columns] = df_points[x_columns] * np.where(orientation >= 0, 1, -1).reshape(-1, 1)
+    try:
+        if data_type == 'video':
+            # print(f"Flipping X coordinates for {data_type} data")
+            
+            # Select all 'x' coordinates
[…]
