+++ Sports2D/detect_pose.py
@@ -230,7 +230,7 @@ def sort_people(keyptpre, keypt, nb_persons_to_detect):
 def save_to_openpose(json_file_path, keypoints, scores, kpt_thr):
     score_threshold = kpt_thr
     nb_detections = len(keypoints)
-    print(f" number of detections: {nb_detections}")
+    # print(f" number of detections: {nb_detections}")
     detections = []
 
     for i in range(nb_detections):
@@ -253,7 +253,7 @@ def save_to_openpose(json_file_path, keypoints, scores, kpt_thr):
                 "hand_left_keypoints_3d": [],
                 "hand_right_keypoints_3d": []
             })
-            print(f"Person {i} passed the threshold. Average score: {average_score:.4f}")
+            # print(f"Person {i} passed the threshold. Average score: {average_score:.4f}")
         # else:
         #     print(f"Person {i} did not pass the threshold. Average score: {average_score:.4f}")
     
@@ -497,8 +497,9 @@ def process_video(video_path, video_result_path,pose_tracker, tracking, output_f
     if display_detection and 'google.colab' not in sys.modules:
         cv2.destroyAllWindows()
 
-def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segment_angles, save_video, save_images, interp_gap_smaller_than, 
-                   filter_options, show_plots, flip_left_right, kpt_thr, data_type, do_filter_angles, min_detection_time):
+def process_webcam(webcam_settings, pose_tracker, openpose_skeleton, joint_angles, segment_angles, 
+                       save_vid, save_img, interp_gap_smaller_than, filter_options, show_plots, flip_left_right,
+                         tracking, kpt_thr, data_type, min_detection_time, filter_options_ang, show_plots_ang):
     """
     Process a live webcam feed to detect poses and calculate angles.
 
@@ -637,7 +638,7 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
         cv2.namedWindow("Real-time Analysis", cv2.WINDOW_NORMAL)
 
     # Image output directory setup
-    if save_images:
+    if save_img:
         img_output_dir = output_dir / 'webcam_images'
         img_output_dir.mkdir(parents=True, exist_ok=True)
     
@@ -669,7 +670,8 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
             
             # Reorder keypoints, scores
             if tracking:
-                max_id = max(pose_tracker.track_ids_last_frame)
+                print("트레킹시작할게요!!!!!")
+                max_id = max(pose_tracker.track_ids_last_frame) if pose_tracker.track_ids_last_frame else 0
                 num_frames, num_points, num_coordinates = keypoints.shape
                 keypoints_filled = np.zeros((max_id+1, num_points, num_coordinates))
                 scores_filled = np.zeros((max_id+1, num_points))
@@ -687,8 +689,8 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
                 person_keypoints = keypoints[person_idx]
                 person_scores = scores[person_idx]
 
-                if np.sum(person_scores >= kpt_thr) < len(person_keypoints) * 0.3:
-                    continue  
+                # if np.sum(person_scores >= kpt_thr) < len(person_keypoints) * 0.3:
+                    # continue  
 
                 df_points = common.convert_keypoints_to_dataframe(person_keypoints, person_scores)
 
@@ -700,14 +702,14 @@ def process_webcam(webcam_settings, pose_tracker, tracking, joint_angles, segmen
                 valid_scores.append(person_scores)
                 valid_person_ids.append(person_idx)
 
-                print(f"Person {person_idx} - Before flip:")
[…]
