+++ Sports2D/compute_angles.py
@@ -230,7 +230,7 @@ def flip_left_right_direction(df_points, data_type):      # ! should check worki
     - df_points: dataframe of pose detection with flipped X coordinates
     '''
 
-    print(f"Flipping left-right direction for {data_type} data")
+    # print(f"Flipping left-right direction for {data_type} data")
     if data_type not in ['video', 'webcam']:
         raise ValueError("data_type must be either 'video' or 'webcam'")
     
@@ -561,8 +561,8 @@ def draw_joint_angle(frame, joint, angle, keypoints, scores, kpt_thr):
         "Left knee": [11, 13, 15],    # left hip, knee, ankle
         "Right hip": [6, 12, 14],    # right shoulder, right hip, right knee
         "Left hip": [5, 11, 13],     # left shoulder, left hip, left knee
-        "Right shoulder": [19, 6, 8], # neck, shoulder, elbow
-        "Left shoulder": [19, 5, 7],  # neck, shoulder, elbow
+        "Right shoulder": [12, 6, 8], # right hip, shoulder, elbow
+        "Left shoulder": [11, 5, 7],  # left hip, shoulder, elbow
         "Right elbow": [6, 8, 10],    # shoulder, elbow, wrist
         "Left elbow": [5, 7, 9],      # shoulder, elbow, wrist
     }
@@ -571,11 +571,13 @@ def draw_joint_angle(frame, joint, angle, keypoints, scores, kpt_thr):
         pts = [keypoints[i] for i in joint_to_keypoints[joint]]
         scores_pts = [scores[i] for i in joint_to_keypoints[joint]]
         if all(score >= kpt_thr for score in scores_pts):
+            neck = keypoints[18]  # neck keypoint
+            mid_hip = keypoints[19]  # mid hip keypoint
             if 'ankle' in joint.lower():
-                draw_angle_arc(frame, joint, pts, angle)
+                draw_angle_arc(frame, joint, pts, angle, neck, mid_hip)
             else:
                 pt1, pt2, pt3 = pts
-                draw_angle_arc(frame, joint, [pt1, pt2, pt3], angle)
+                draw_angle_arc(frame, joint, [pt1, pt2, pt3], angle, neck, mid_hip)
 
 def draw_dotted_line(frame, start, direction, length, color=(0, 255, 0), gap=7, dot_length=3): # default is green color
     for i in range(0, length, gap):
@@ -584,7 +586,7 @@ def draw_dotted_line(frame, start, direction, length, color=(0, 255, 0), gap=7, 
         cv2.line(frame, tuple(line_start.astype(int)), tuple(line_end.astype(int)), color, 2)
 
 # everything well done.
-def draw_angle_arc(frame, joint, pts, angle):
+def draw_angle_arc(frame, joint, pts, angle, neck, mid_hip):
     """
     """
     start_angle = 0
@@ -601,13 +603,10 @@ def draw_angle_arc(frame, joint, pts, angle):
             # Calculate shank vector (from ankle to knee)
             shank_vec = knee - ankle
             
-            # Calculate foot vector (from ankle to toe)
-            foot_vec = toe - ankle
+            # Calculate foot vector (from heel to toe)
+            foot_vec = toe - heel
 
-            # Vector length
-            vector_length = np.linalg.norm(shank_vec)
-            
-            # Calculate perpendicular vector to shank
+            # Calculate perpendicular vector to shank (shin)
             perpendicular_vec = np.array([-shank_vec[1], shank_vec[0]])
             perpendicular_vec /= np.linalg.norm(perpendicular_vec)
             
@@ -618,44 +617,52 @@ def draw_angle_arc(frame, joint, pts, angle):
             radius = int(0.2 * np.linalg.norm(knee - ankle))
             ref_point = ankle
             
[…]
