+++ Sports2D/compute_angles.py
@@ -84,21 +84,6 @@ __status__ = "Development"
 # text dics
 joint_text_positions = {}
 
-# segment와 joint 매핑 정의
-segment_to_joint = {
-    "Right foot": "Right ankle",
-    "Left foot": "Left ankle",
-    "Right shank": "Right knee",
-    "Left shank": "Left knee",
-    "Right thigh": "Right hip",
-    "Left thigh": "Left hip",
-    "Trunk": None,  # Trunk에 해당하는 단일 joint가 없음
-    "Right arm": "Right shoulder",
-    "Left arm": "Left shoulder",
-    "Right forearm": "Right elbow",
-    "Left forearm": "Left elbow"
-}
-
 # define joint angle parameters
 def get_joint_angle_params(joint):
     """
@@ -142,7 +127,7 @@ def get_segment_angle_params(segment):
         'Left shank': [['left_knee', 'left_ankle'], 'horizontal', 0, -1],
         'Right thigh': [['right_hip', 'right_knee'], 'horizontal', 0, -1],
         'Left thigh': [['left_hip', 'left_knee'], 'horizontal', 0, -1],
-        'Trunk': [['right_hip', 'right_shoulder'], 'horizontal', 0, -1],
+        'Trunk': [['hip', 'neck'], 'horizontal', 0, -1],
         'Right arm': [['right_shoulder', 'right_elbow'], 'horizontal', 0, -1],
         'Left arm': [['left_shoulder', 'left_elbow'], 'horizontal', 0, -1],
         'Right forearm': [['right_elbow', 'right_wrist'], 'horizontal', 0, -1],
@@ -522,61 +507,20 @@ def segment_angles_series_from_csv(df_points, angle_params, segment, kpt_thr):
         
     return ang_series
 
-# def adjust_text_scale(frame, base_scale=0.25, base_thickness=1):
-#     """
-#     Adjusts the text scale and thickness based on the frame size.
-
-#     This function calculates appropriate text scale and thickness for overlay text
-#     on a given frame. It scales the text properties relative to a base frame width
-#     of 640 pixels, ensuring that text appears proportional on different frame sizes.
-
-#     Args:
-#         frame (numpy.ndarray): The input frame (image) on which text will be drawn.
-#                                Expected to be a 3-dimensional array (height, width, channels).
-#         base_scale (float, optional): The base scale factor for text. Defaults to 0.25.
-#         base_thickness (int, optional): The base thickness for text. Defaults to 1.
-
-#     Returns:
-#         tuple: A tuple containing two elements:
-#                - scale (float): The adjusted scale factor for text.
-#                - thickness (int): The adjusted thickness for text.
-#     """
-#     height, width, _ = frame.shape
-#     scale = base_scale * (width / 640)
-#     thickness = int(base_thickness * (width / 640))
-#     return scale, thickness
-
-def calculate_dynamic_font_scale(vector_length, min_scale=0.2, max_scale=3.0):
-    # 벡터 길이의 20%를 기준으로 폰트 크기 계산
-    base_font_scale = vector_length * 0.005  # 0.005는 조정 가능한 계수
-    
-    # 최소/최대 스케일 적용
-    return max(min(base_font_scale, max_scale), min_scale)
-
 def draw_joint_angle(frame, joint, angle, keypoints, scores, kpt_thr):
-    joint_to_keypoints = {
-        "Right ankle": [14, 16, 25, 21],  # right knee, right ankle, right heel, right big_toe
-        "Left ankle": [13, 15, 24, 20],   # left knee, left ankle, left heel, left big_toe
-        "Right knee": [12, 14, 16],   # right hip, right knee, right ankle
[…]
