+++ Sports2D/detect_pose.py
@@ -234,10 +234,10 @@ def save_to_openpose(json_file_path, keypoints, scores, kpt_thr):
     detections = []
 
     for i in range(nb_detections):
-        # 각 사람의 평균 score 계산
+        # average score of all keypoints
         average_score = sum(score.item() for score in scores[i]) / len(scores[i])
         
-        # 평균 score가 임계값 이상인 경우만 처리
+        # if average score is above threshold, add to detections
         if average_score >= score_threshold:
             keypoints_with_confidence_i = []
             for kp, score in zip(keypoints[i], scores[i]):
@@ -289,14 +289,12 @@ def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, 
     keypoints_ids = [kp['id'] for kp in keypoints_info.values()]
     keypoints_names = [kp['name'] for kp in keypoints_info.values()]
     keypoints_nb = len(keypoints_ids)
-    print(f"keypoints_nb: {keypoints_nb}")
 
     # Retrieve coordinates
     logging.info('Sorting people across frames.')
     json_fnames = sorted(list(json_path.glob('*.json')))
     # print(f"json_fnames: {json_fnames}")
     nb_persons_to_detect = max([len(json.load(open(json_fname))['people']) for json_fname in json_fnames])
-    print(f"nb_persons_to_detect: {nb_persons_to_detect}")
     Coords = [np.array([]).reshape(0,keypoints_nb*3)] * nb_persons_to_detect
     for json_fname in json_fnames:    # for each frame
         with open(json_fname) as json_f:
@@ -396,7 +394,7 @@ def show_image(img, title):
     return True
     
 # If input is a video
-def process_video(video_path, video_result_path,pose_tracker, tracking, output_format, save_video, save_images, display_detection, frame_range, kpt_thr):
+def process_video(video_path, video_result_path,pose_tracker, output_format, save_video, save_images, display_detection, frame_range, kpt_thr):
     '''
     Estimate pose from a video file
     
@@ -425,7 +423,6 @@ def process_video(video_path, video_result_path,pose_tracker, tracking, output_f
         raise NameError(f"{video_path} is not a video. Images must be put in one subdirectory per camera.")
     
     pose_dir = os.path.abspath(os.path.join(video_result_path, '..', 'video_results'))
-    print(f"pose_dir: {pose_dir}")
     if not os.path.isdir(pose_dir): os.makedirs(pose_dir)
     video_name_wo_ext = os.path.splitext(os.path.basename(video_path))[0]
     json_output_dir = os.path.join(pose_dir, f'{video_name_wo_ext}_json')
@@ -452,17 +449,6 @@ def process_video(video_path, video_result_path,pose_tracker, tracking, output_f
             if frame_idx in range(*f_range):
                 # Perform pose estimation on the frame
                 keypoints, scores = pose_tracker(frame)
-
-                # Reorder keypoints, scores
-                if tracking:
-                    max_id = max(pose_tracker.track_ids_last_frame)
-                    num_frames, num_points, num_coordinates = keypoints.shape
-                    keypoints_filled = np.zeros((max_id+1, num_points, num_coordinates))
-                    scores_filled = np.zeros((max_id+1, num_points))
-                    keypoints_filled[pose_tracker.track_ids_last_frame] = keypoints
-                    scores_filled[pose_tracker.track_ids_last_frame] = scores
-                    keypoints = keypoints_filled
-                    scores = scores_filled
 
                 # Save to json
                 if 'openpose' in output_format:
@@ -499,7 +485,7 @@ def process_video(video_path, video_result_path,pose_tracker, tracking, output_f
 
[…]
