+++ Sports2D/detect_pose.py
@@ -228,6 +228,19 @@ def sort_people(keyptpre, keypt, nb_persons_to_detect):
     return keypt
 
 def save_to_openpose(json_file_path, keypoints, scores, kpt_thr):
+    """
+    Saves the detected keypoints and scores to a JSON file in the OpenPose format.
+
+    Args:
+        json_file_path (str): The path to the output JSON file.
+        keypoints (list): A list of detected keypoints, where each element is a list of 2D coordinates for each keypoint.
+        scores (list): A list of scores for each detected keypoint, where each element is a list of confidence scores for each keypoint.
+        kpt_thr (float): The minimum average score threshold for a detection to be included in the output.
+
+    Returns:
+        None
+    """
+
     score_threshold = kpt_thr
     nb_detections = len(keypoints)
     # print(f" number of detections: {nb_detections}")
@@ -243,7 +256,7 @@ def save_to_openpose(json_file_path, keypoints, scores, kpt_thr):
             for kp, score in zip(keypoints[i], scores[i]):
                 keypoints_with_confidence_i.extend([kp[0].item(), kp[1].item(), score.item()])
             detections.append({
-                "person_id": [-1],
+                "person_id": [-1], # Maybe change to i ?
                 "pose_keypoints_2d": keypoints_with_confidence_i,
                 "face_keypoints_2d": [],
                 "hand_left_keypoints_2d": [],
@@ -278,6 +291,7 @@ def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, 
     - filter_options: list, options for filtering
     - show_plots: boolean, show plots or not
     - min_detection_time: minimum detection time in seconds (default: 1 second)
+    - tracking: boolean, use tracking or not (when input_data is video)
 
     OUTPUTS:
     - Creation of one csv files per detected person
@@ -293,7 +307,6 @@ def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, 
     # Retrieve coordinates
     logging.info('Sorting people across frames.')
     json_fnames = sorted(list(json_path.glob('*.json')))
-    # print(f"json_fnames: {json_fnames}")
     nb_persons_to_detect = max([len(json.load(open(json_fname))['people']) for json_fname in json_fnames])
     Coords = [np.array([]).reshape(0,keypoints_nb*3)] * nb_persons_to_detect
     for json_fname in json_fnames:    # for each frame
@@ -313,7 +326,6 @@ def json_to_csv(json_path, frame_rate, interp_gap_smaller_than, filter_options, 
                 keyptpre = keypt
             # Associate persons across frames if tracking is enabled
             if tracking and keyptpre is not None:
-                # print(f"Tracking enabled. keyptpre: {keyptpre.shape}")
                 keypt = sort_people(keyptpre, keypt, nb_persons_to_detect)
             # Concatenate to coordinates of previous frames
             for i in range(nb_persons_to_detect): 
@@ -403,12 +415,12 @@ def process_video(video_path, video_result_path,pose_tracker, output_format, sav
     INPUTS:
     - video_path: str. Path to the input video file
     - pose_tracker: PoseTracker. Initialized pose tracker object from RTMLib
-    - tracking: bool. Whether to give consistent person ID across frames
     - output_format: str. Output format for the pose estimation results ('openpose', 'mmpose', 'deeplabcut')
     - save_video: bool. Whether to save the output video
     - save_images: bool. Whether to save the output images
     - display_detection: bool. Whether to show real-time visualization
     - frame_range: list. Range of frames to process
+    - kpt_thr: float. Keypoint score threshold
 
     OUTPUTS:
     - JSON files with the detected keypoints and confidence scores in the OpenPose format
[…]
