+++ Sports2D/compute_angles.py
@@ -203,7 +203,7 @@ def points2D_to_angles(points_list):
 
     return ang_deg
 
-def flip_left_right_direction(df_points, data_type):      # ! should check working on video and webcam (viedo ok, webcam not checked)
+def flip_left_right_direction(df_points, data_type):
     '''
     Inverts X coordinates to get consistent angles when person changes direction and goes to the left.
     The person is deemed to go to the left when their toes are to the left of their heels.
@@ -215,13 +215,8 @@ def flip_left_right_direction(df_points, data_type):      # ! should check worki
     - df_points: dataframe of pose detection with flipped X coordinates
     '''
 
-    # print(f"Flipping left-right direction for {data_type} data")
     if data_type not in ['video', 'webcam']:
         raise ValueError("data_type must be either 'video' or 'webcam'")
-    
-    # print('Flipping left-right direction...')
-    # print("DataFrame columns:", df_points.columns)
-    # print("Index levels:", df_points.columns.names)
     
     # Set the level names explicitly if they are None
     if df_points.columns.names == [None, None, None, None]:
@@ -229,12 +224,9 @@ def flip_left_right_direction(df_points, data_type):      # ! should check worki
     
     try:
         if data_type == 'video':
-            # print(f"Flipping X coordinates for {data_type} data")
             
             # Select all 'x' coordinates
             x_coords = df_points.xs('x', axis=1, level='coords')
-            # print(f"X coordinates first 5 values:\n{x_coords.head()}")
-            # print(f"X coordinates columns: {x_coords.columns}")
             
             # Access MultiIndex columns correctly
             scorer = x_coords.columns.get_level_values('scorer')[0]
@@ -259,17 +251,10 @@ def flip_left_right_direction(df_points, data_type):      # ! should check worki
             left_toe = df_points['left_big_toe_x']
             left_heel = df_points['left_heel_x']
         
-        # print(f"Right toe first 5 values: {right_toe.head()}")
-        # print(f"Right heel first 5 values: {right_heel.head()}")
-        # print(f"Left toe first 5 values: {left_toe.head()}")
-        # print(f"Left heel first 5 values: {left_heel.head()}")
-        
         # Calculate orientation while handling NaN values
         right_orientation = right_toe - right_heel
         left_orientation = left_toe - left_heel
         orientation = right_orientation.add(left_orientation, fill_value=0)
-
-        # print(f"Orientation first 5 values: {orientation.head()}")
         
         # Create flip mask, treating NaN as positive (no flip)
         flip_mask = np.where(orientation.isnull(), 1, np.where(orientation >= 0, 1, -1))
@@ -280,7 +265,6 @@ def flip_left_right_direction(df_points, data_type):      # ! should check worki
             
             # Update the original dataframe with flipped x coordinates
             df_points.loc[:, (slice(None), slice(None), slice(None), 'x')] = x_coords_flipped.values
-            # print(f"Flipped X coordinates first 5 values:\n{x_coords_flipped.head()}")
 
         else:  # webcam
             x_columns = [col for col in df_points.columns if col.endswith('_x')]
@@ -404,7 +388,6 @@ def joint_angles_series_from_csv(df_points, angle_params, kpt_thr):
     OUTPUT:
     - ang_series: array of time series of the considered angle
     '''
-    print(f"Processing angle for: {angle_params[0]}")
     
     # Retrieve points
[…]
