+++ Sports2D/Demo/Config_demo.toml
@@ -14,7 +14,7 @@
 
 [project]
 video_dir = '' # BETWEEN SINGLE QUOTES! # If empty, result dir is current dir
-video_input = 'demo.mp4' # 'webcam' or 'your_video_name.extenstion'
+video_input = 'webcam' # 'webcam' or 'your_video_name.extenstion'
 result_dir = '' # BETWEEN SINGLE QUOTES! # If empty, project dir is current dir
 
 [pose]
