+++ Sports2D/detect_pose.py
@@ -42,6 +42,7 @@
 ## INIT
 import os
 import logging
+import time as time_module # Avoid conflict
 from pathlib import Path
 from sys import platform
 import json
@@ -56,7 +57,6 @@ import torch
 import onnxruntime as ort
 from datetime import datetime
 import sys
-import time
 from IPython.display import clear_output
 
 from Sports2D.Sports2D import base_params
@@ -485,6 +485,8 @@ def process_video(video_path, video_result_path,pose_tracker, output_format, sav
     if display_detection and 'google.colab' not in sys.modules:
         cv2.destroyAllWindows()
 
+
+import time
 def process_webcam(cam_id, pose_tracker, openpose_skeleton, joint_angles, segment_angles, 
                        save_vid, save_img, interp_gap_smaller_than, filter_options, show_plots, flip_left_right,
                          kpt_thr, data_type, min_detection_time, filter_options_ang, show_plots_ang):
@@ -531,17 +533,154 @@ def process_webcam(cam_id, pose_tracker, openpose_skeleton, joint_angles, segmen
     Note: The function can be interrupted by pressing 'q' or using a keyboard interrupt (Ctrl+C).
     """
 
-    global time
-
     is_colab = 'google.colab' in sys.modules
     if is_colab:
         from IPython.display import display, Javascript, clear_output
         from google.colab.output import eval_js
-        from base64 import b64decode
+        from base64 import b64decode, b64encode
         import matplotlib.pyplot as plt
         import PIL.Image
         import io
+    
+        # JavaScript to properly create our live video stream using our webcam as input
+        def video_stream():
+            js = Javascript('''
+                var video;
+                var div = null;
+                var stream;
+                var captureCanvas;
+                var imgElement;
+                var labelElement;
+                var resultImgElement;
+                
+                var pendingResolve = null;
+                var shutdown = false;
+                
+                function removeDom() {
+                   stream.getVideoTracks()[0].stop();
+                   video.remove();
+                   div.remove();
+                   video = null;
+                   div = null;
+                   stream = null;
+                   imgElement = null;
+                   captureCanvas = null;
+                   labelElement = null;
+                   resultImgElement = null;
+                }
+                
+                function onAnimationFrame() {
+                  if (!shutdown) {
+                    window.requestAnimationFrame(onAnimationFrame);
+                  }
+                  if (pendingResolve) {
+                    var result = "";
+                    if (!shutdown) {
+                      captureCanvas.getContext('2d').drawImage(video, 0, 0, 640, 480);
+                      result = captureCanvas.toDataURL('image/jpeg', 0.8)
+                    }
+                    var lp = pendingResolve;
+                    pendingResolve = null;
+                    lp(result);
+                  }
+                }
+                
+                async function createDom() {
+                  if (div !== null) {
+                    return stream;
+                  }
+            
+                  div = document.createElement('div');
+                  div.style.border = '2px solid black';
+                  div.style.padding = '3px';
+                  div.style.width = '100%';
+                  div.style.maxWidth = '600px';
+                  document.body.appendChild(div);
+                  
+                  const modelOut = document.createElement('div');
+                  modelOut.innerHTML = "<span>Status:</span>";
+                  labelElement = document.createElement('span');
+                  labelElement.innerText = 'No data';
+                  labelElement.style.fontWeight = 'bold';
+                  modelOut.appendChild(labelElement);
+                  div.appendChild(modelOut);
+                       
[…]
