### 2024-08-06 d80f39f
remove torch in dependencies
