### 2024-08-06 #6849 @Kevin-Mattheus-Moerman
@tuliofalmeida it looks like you've ticked all boxes. Are you happy to recommend acceptance of this work for JOSS? Thanks for your help!
