### 2024-08-07 0680760
fix bug of frame_range(time_range)
