+++ Sports2D/detect_pose.py
@@ -1057,7 +1057,7 @@ def detect_pose_fun(config_dict, video_file):
     # Pose settings
     time_range = config_dict.get('pose', {}).get('time_range', []) 
     frame_range = [int(time_range[0] * frame_rate), int(time_range[1] * frame_rate)] if len(time_range) == 2 else []
-    display_detection = config_dict.get('pose_advanced').get('display_detection')
+    display_detection = config_dict.get('pose').get('display_detection')
     output_format = "openpose"
     openpose_skeleton = False
     
