+++ Sports2D/compute_angles.py
@@ -998,7 +998,7 @@ def draw_keypts_skel(X, Y, scores, img, pose_model, kpt_thr):
     
     return img
 
-def save_imgvid_reID(video_path, video_result_path, df_angles_list, pose_model, save_vid, save_img, kpt_thr):
+def save_imgvid_reID(video_path, video_result_path, df_angles_list, pose_model, save_vid, save_img, kpt_thr, frame_range):
     """
     Saves the video and/or images with the computed angles and keypoints overlaid on the original video.
 
@@ -1010,6 +1010,7 @@ def save_imgvid_reID(video_path, video_result_path, df_angles_list, pose_model, 
         save_vid (bool): Whether to save the video with the overlaid angles and keypoints.
         save_img (bool): Whether to save individual frames as images with the overlaid angles and keypoints.
         kpt_thr (float): Threshold for displaying keypoints.
+        frame_range (list): Range of frames to process. If empty, process all frames.
 
     Returns:
         None
@@ -1021,7 +1022,17 @@ def save_imgvid_reID(video_path, video_result_path, df_angles_list, pose_model, 
     if not csv_paths:
         print("Error: No CSV files found in the specified directory.")
         return
-        
+    
+    cap = cv2.VideoCapture(str(video_path))
+    if not cap.isOpened():
+        print(f"Error opening video file: {video_path}")
+        return
+
+    total_frames = int(cap.get(cv2.CAP_PROP_FRAME_COUNT))
+    f_range = [0, total_frames] if frame_range == [] else frame_range
+
+    print(f"Processing frames from {f_range[0]} to {f_range[1]}")
+
     # Load both angles and points CSV files
     angles_coords = []
     points_coords = []
@@ -1040,10 +1051,7 @@ def save_imgvid_reID(video_path, video_result_path, df_angles_list, pose_model, 
             person_id = int(c.stem.split('_person')[1].split('_')[0])
             person_ids.append(person_id)
 
-    cap = cv2.VideoCapture(str(video_path))
-    if not cap.isOpened():
-        print(f"Error opening video file: {video_path}")
-        return
+    print(f"Loaded data for {len(person_ids)} persons")
 
     W, H = int(cap.get(cv2.CAP_PROP_FRAME_WIDTH)), int(cap.get(cv2.CAP_PROP_FRAME_HEIGHT))
     fps = cap.get(cv2.CAP_PROP_FPS)
@@ -1057,30 +1065,24 @@ def save_imgvid_reID(video_path, video_result_path, df_angles_list, pose_model, 
         img_pose_path = csv_dir / (video_result_path.stem + '_' + pose_model + '_img')
         img_pose_path.mkdir(parents=True, exist_ok=True)
         
-    f = 0
-    while(cap.isOpened()):
+    cap.set(cv2.CAP_PROP_POS_FRAMES, f_range[0])
+    for f in range(len(points_coords[0])): # csv file already has the same number of frames
         ret, frame = cap.read()
         if not ret:
+            print(f"Failed to read frame {f + f_range[0]}")
             break
         
         frame_keypoints = []
         frame_scores = []
         for coord in points_coords:
-            if f < len(coord):
-                X = np.array(coord.iloc[f, 0::3])
-                Y = np.array(coord.iloc[f, 1::3])
-                S = np.array(coord.iloc[f, 2::3])
-                
-                person_keypoints = np.column_stack((X, Y))
-                frame_keypoints.append(person_keypoints)
-                frame_scores.append(S)
-            else:
-                if frame_keypoints and frame_scores:
-                    frame_keypoints.append(frame_keypoints[-1])
-                    frame_scores.append(frame_scores[-1])
-                else:
[…]
