+++ Sports2D/Sports2D.py
@@ -244,25 +244,3 @@ def compute_angles(config='Config_demo.toml'):
         
         end = time.time()
         logging.info(f'Joint and segment computation took {end-start:.2f} s.')
-        
- 
-def run_rtmpose(config='Config_demo.toml'):
-    '''
-    Run RTMPose for pose estimation and angle computation.
-    '''
-    config_dict = read_config_file(config)
-    _, video_files, result_dir, _ = base_params(config_dict)
-    
-    with open(result_dir / 'logs.txt', 'a+') as log_f: pass
-    logging.basicConfig(format='%(message)s', level=logging.INFO, force=True, 
-        handlers = [logging.handlers.TimedRotatingFileHandler(result_dir / 'logs.txt', when='D', interval=7), logging.StreamHandler()])
-    
-    logging.info("\n\n---------------------------------------------------------------------")
-    logging.info("Running RTMPose for pose estimation and angle computation")
-    logging.info("---------------------------------------------------------------------")
-    start = time.time()
-    
-    rtm_estimator(config_dict)
-    
-    end = time.time()
-    logging.info(f'RTMPose estimation and angle computation took {end-start:.2f} s.')
