### 2024-08-07 a4a3847
Unify Calculation functions
