+++ Sports2D/compute_angles.py
@@ -276,219 +276,112 @@ def flip_left_right_direction(df_points, data_type):
 
     return df_points
 
-def joint_angles_series_from_points(df_points, angle_params, kpt_thr):
+def calculate_angles_series(df_points, angle_params, segment=None, kpt_thr=0.0, is_csv=False):
     '''
-    Obtain joint angle series from point series.
+    Obtain joint or segment angle series from point series.
     
     INPUT: 
-    - df_points: dataframe of pose detection, from csv
+    - df_points: dataframe of pose detection
     - angle_params: dictionary specifying which points to use, and what offset and multiplying factor to use
-    
-    OUTPUT:
-    - ang_series: array of time series of the considered angle
-    '''
-    
-    # Retrieve points
-    keypt_series = []
-    for k in angle_params[0]:
-        if f"{k}_x" in df_points.columns and f"{k}_y" in df_points.columns and f"{k}_score" in df_points.columns:
-            score = df_points[f"{k}_score"].values[0]
-            if score >= kpt_thr:
-                keypt = df_points[[f"{k}_x", f"{k}_y"]]
-                keypt_series.append(keypt)
-            else:
-                return None
-        else:
-            print(f"Error: Keypoint {k} or its score not found in dataframe")
-            return None
-
-    if len(keypt_series) != len(angle_params[0]):
-        print(f"Error: Not all required keypoints were found or passed the threshold")
-        return None
-
-    # Compute angles
-    points_list = [k.values.T for k in keypt_series]
-    ang_series = points2D_to_angles(points_list)
-    if ang_series is None:
-        print(f"points2D_to_angles returned None for {angle_params[1]}")
-        return None
-    
-    ang_series += angle_params[2]
-    ang_series *= angle_params[3]
-    
-    if ang_series.mean() > 180: ang_series -= 360
-    if ang_series.mean() < -180: ang_series += 360
-
-    return ang_series
-
-
-def segment_angles_series_from_points(df_points, angle_params, segment, kpt_thr):
-    '''
-    Obtain segment angle series w/r horizontal from point series.
-    For trunk segment: mean of the angles between RHip-RShoulder and LHip-LShoulder
-    
-    INPUT: 
-    - df_points: dataframe of pose detection, from csv
-    - angle_params: dictionary specifying which points to use, and what offset and multiplying factor to use
-    - segment: which segment angle is considered
-    
-    OUTPUT:
-    - ang_series: array of time series of the considered angle
-    '''
-    
-    # Retrieve points
-    keypt_series = []
-    for k in angle_params[0]:
-        if df_points[f"{k}_score"].values[0] >= kpt_thr:
-            keypt_series.append(df_points[[f"{k}_x", f"{k}_y"]])
-        else:
-            return None
-
-    points_list = [k.values.T for k in keypt_series]
-    ang_series = points2D_to_angles(points_list)
-    if ang_series is None:
-        return None
-
-    ang_series += angle_params[2]
-    ang_series *= angle_params[3]
-    # ang_series = np.where(ang_series>180,ang_series-360,ang_series)
-
-    # For trunk: mean between angles RHip-RShoulder and LHip-LShoulder
-    if segment == 'Trunk':
-        ang_seriesR = ang_series
-        angle_params[0] = [a.replace('R','L') for a in angle_params[0]]
-        keypt_series = []
-        for k in angle_params[0]:
-            if df_points[f"{k}_score"].values[0] >= kpt_thr:
-                keypt_series.append(df_points[[f"{k}_x", f"{k}_y"]])
-            else:
-                return None
-
-        points_list = [k.values.T for k in keypt_series]
[…]
