### 2024-08-07 ce0cd1f
fix time_range error when using webcam
