+++ setup.cfg
@@ -44,7 +44,6 @@ install_requires =
     rtmlib
     openvino
     tqdm
-    onnxruntime
 packages = find:
 
 [options.package_data]
