### 2024-08-08 598f79b
add webcam's input_size parameter
