### 2024-08-08 e7a5664
add webcam's input_size parameter
