+++ Sports2D/Utilities/test_with_RTMPose.toml
@@ -1,0 +1,77 @@
+###############################################################################
+## SPORTS2D PROJECT PARAMETERS                                               ##
+###############################################################################
+
+# Configure your project parameters here
+
+# Then open an Anaconda prompt and enter:
+# conda activate Sports2D
+# ipython
+# from Sports2D import Sports2D
+# Sports2D.detect_pose('Config_demo.toml')
+# Sports2D.compute_angles('Config_demo.toml')
+
+
+[project]
+video_dir = '' # BETWEEN SINGLE QUOTES! # If empty, result dir is current dir
+video_input = 'demo.mp4' # 'webcam' or 'your_video_name.extenstion'
+result_dir = '' # BETWEEN SINGLE QUOTES! # If empty, project dir is current dir
+
+[pose]
+display_detection = false
+time_range = []
+
+[compute_angles]
+# Select joint angles among
+# ['Right ankle', 'Left ankle', 'Right knee', 'Left knee', 'Right hip', 'Left hip', 'Right shoulder', 'Left shoulder', 'Right elbow', 'Left elbow', 'Right wrist', 'Left wrist']
+joint_angles = ['Right ankle', 'Left ankle', 'Right knee', 'Left knee', 'Right hip', 'Left hip', 'Right shoulder', 'Left shoulder', 'Right elbow', 'Left elbow']
+
+# Select segment angles among
+# ['Right foot', 'Left foot', 'Right shank', 'Left shank', 'Right thigh', 'Left thigh', 'Trunk', 'Right arm', 'Left arm', 'Right forearm', 'Left forearm', 'Right hand', 'Left hand']
+segment_angles = ['Right foot', 'Left foot', 'Right shank', 'Left shank', 'Right thigh', 'Left thigh', 'Trunk', 'Right arm', 'Left arm', 'Right forearm', 'Left forearm']
+
+   	
+# ADVANCED CONFIGURATION
+	
+[pose_advanced]
+webcam_id = 0 # your webcam id (0 is default)
+input_size = 'auto' # 'auto' or [W, H]. Full image resolution: 'auto'. Custom resolution: list of two integers. Lower resolution will be faster but less precise.
+overwrite_pose = true # If false, don't run openpose again if json pose files are found.
+det_frequency = 10 # detect person every det_frequency (1 = 1 frame)
+mode = "lightweight" # # lightweight, balanced, performance
+keypoints_threshold = 0.4 #If only part of a person is on screen, increase this number to ensure that only correctly detected keypoints are used.
+min_detection_time = 1 # If lower than this, person will be ignored(sec)
+						#For webcams, it is possible to detect the wrong person if the person is only partially detected (usually less than a second).
+						
+save_vid = true
+save_img = true
+interp_gap_smaller_than = 5 # do not interpolate bigger gaps
+filter = true
+show_plots = true
+filter_type = 'butterworth' # butterworth, gaussian, LOESS, median
+   [pose_advanced.butterworth]
+   order = 4 
+   cut_off_frequency = 6 # Hz
+   [pose_advanced.gaussian]
+   sigma_kernel = 1 #px
+   [pose_advanced.loess]
+   nb_values_used = 5 # = fraction of data used * nb frames
+   [pose_advanced.median]
+   kernel_size = 3
+
+[compute_angles_advanced]
+show_angles_on_img = true
+show_angles_on_vid = true
+filter = true
+show_plots = true
+flip_left_right = true # Same angles whether the participant faces left/right. Set it to false if you want timeseries to be continuous even when the participent switches their stance.
+filter_type = 'butterworth' # butterworth, gaussian, LOESS, median
[…]
