### 2024-08-09 f48c5c0
Fix bug not working multi-person in webcam when using sort_people
