### 2024-08-10 6150b9a
Fix sync between video and overlay
