### 2024-08-19 #6849 @davidpagnon
Hi @Kevin-Mattheus-Moerman,
Thank you for following up. Can we wait about a week before acceptance? We made it work in real-time, and we would like to add one or two lines in the paper to mention it. Also, there is still one point that @nicos1993 brought up that has not been taken into account.
