### 2024-08-19 #6849 @tuliofalmeida
> @tuliofalmeida it looks like you've ticked all boxes. Are you happy to recommend acceptance of this work for JOSS? Thanks for your help!

I'm sorry, I thought I had already answered you.
And yes, I recommend it for publication. It's working well, I reported him a small bug that I found and I also agree with the suggestion that Nico made.
