### 2024-08-20 #6849 @nicos1993
> Hi Nicos,
>
> Thank you for your review!
>
> 1. I fully agree with your statement that the Demo video was probably not the best, given that it does not lie in the plane. I will replace it with a video of someone doing jumping jacks (frontal plane) and another person doing tuck jumps (sagittal plane).
> 2. I believe the joint angle convention is the same as the one from Winter 2009, but I agree that the way it is represented makes it confusing. I will add this figure to the article and the tutorial.
>    ![joint_convention](https://private-user-images.githubusercontent.com/54667644/351387489-8d0a14dd-ca9e-4f3a-b406-d3fe78963c30.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MjQxNzQwMzAsIm5iZiI6MTcyNDE3MzczMCwicGF0aCI6Ii81NDY2NzY0NC8zNTEzODc0ODktOGQwYTE0ZGQtY2E5ZS00ZjNhLWI0MDYtZDNmZTc4OTYzYzMwLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA4MjAlMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwODIwVDE3MDg1MFomWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTFmZTI4NjllMmU4MjY0MjA0NzQ4YTAwMjI1ZGIwNmY4YWUxYTVlZDlkNGUxNTQ4ZTU1ODI5ODE5NjBhNjNlMTMmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.qpm5aS2eKpD43wS1QNa2KRZKBYjcebjw434W4vB6E9s)
>
> I will ping you once the fixes have been made.

Hello @Kevin-Mattheus-Moerman ,

I believe the changes I suggested regarding the examples in the article and paper have not been made/updated yet (please correct me if I am wrong. I also believe the updated joint angle convention figure has not been included in the article.

Best wishes,

Nicos
