### 2024-08-20 #6849 @davidpagnon
You are right, there is more cleaning up to do in the new code than I expected so it takes a bit of time. I'm in a rushed period so it's hard to prioritize it but I'll try to get it done by next week, sorry about it!
