+++ Content/paper.bib
@@ -97,6 +97,19 @@
   publisher={SpringerOpen}
 }
 
+@article{Delp_2007,
+  title={OpenSim: open-source software to create and analyze dynamic simulations of movement},
+  author={Delp, Scott L and Anderson, Frank C and Arnold, Allison S and Loan, Peter and Habib, Ayman and John, Chand T and Guendelman, Eran and Thelen, Darryl G},
+  journal={IEEE transactions on biomedical engineering},
+  volume={54},
+  number={11},
+  pages={1940--1950},
+  year={2007},
+  URL = {https://ieeexplore.ieee.org/abstract/document/4352056},
+  DOI = {10.1109/TBME.2007.901024},
+  publisher={IEEE}
+}
+
 @article{Di_2021,
   title={A 2D qualitative movement assessment of a deceleration task detects football players with high knee joint loading},
   author={Di Paolo, Stefano and Zaffagnini, Stefano and Tosarelli, Filippo and Aggio, Fabrizio and Bragonzoni, Laura and Grassi, Alberto and Della Villa, Francesco},
@@ -106,6 +119,18 @@
   pages={4032--4040},
   year={2021},
   publisher={Springer}
+}
+
+@article{Jiang_2023,
+  title={RTMPose: Real-Time Multi-Person Pose Estimation based on MMPose}, 
+  author={Tao Jiang and Peng Lu and Li Zhang and Ningsheng Ma and Rui Han and Chengqi Lyu and Yining Li and Kai Chen},
+  DOI={10.48550/arXiv.2303.07399},
+  journal={arXiv},
+  year={2023},
+  eprint={2303.07399},
+  archivePrefix={arXiv},
+  primaryClass={cs.CV},
+  url={https://arxiv.org/abs/2303.07399}, 
 }
 
 @article{Kidzinski_2020,
@@ -198,6 +223,20 @@
   pages={155--162},
   year={2016},
   publisher={Taylor \& Francis}
+}
+
+@article{Seth_2018,
+  DOI = {10.1371/journal.pcbi.1006223},
+  author = {Seth, Ajay AND Hicks, Jennifer L. AND Uchida, Thomas K. AND Habib, Ayman AND Dembia, Christopher L. AND Dunne, James J. AND Ong, Carmichael F. AND DeMers, Matthew S. AND Rajagopal, Apoorva AND Millard, Matthew AND Hamner, Samuel R. AND Arnold, Edith M. AND Yong, Jennifer R. AND Lakshmikanth, Shrinidhi K. AND Sherman, Michael A. AND Ku, Joy P. AND Delp, Scott L.},
+  journal = {PLOS Computational Biology},
+  publisher = {Public Library of Science},
+  title = {OpenSim: Simulating musculoskeletal dynamics and neuromuscular control to study human and animal movement},
+  year = {2018},
+  month = {07},
+  volume = {14},
+  url = {https://doi.org/10.1371/journal.pcbi.1006223},
+  pages = {1-20},
+  number = {7},
 }
 
  @article{Uhlrich_2022,
