+++ Sports2D/detect_pose.py
@@ -1,661 +1,0 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-
-'''
-    ##############################################################
-    ## Compute angles from 2D pose detection                    ##
-    ##############################################################
-    
-    Detect joint centers from a video with OpenPose or BlazePose.
-    Save a 2D csv position file per person, and optionally json files, image files, and video files.
-    
-    If OpenPose is used, multiple persons can be consistently detected across frames.
-    Interpolates sequences of missing data if they are less than N frames long.
-    Optionally filters results with Butterworth, gaussian, median, or loess filter.
-    Optionally displays figures.
-
-    If BlazePose is used, only one person can be detected.
-    No interpolation nor filtering options available. Not plotting available.
-
-    /!\ Warning /!\
-    - The pose detection is only as good as the pose estimation algorithm, i.e., it is not perfect.
-    - It will lead to reliable results only if the persons move in the 2D plane (sagittal plane).
-    - The persons need to be filmed as perpendicularly as possible from their side.
-    If you need research-grade markerless joint kinematics, consider using several cameras,
-    and constraining angles to a biomechanically accurate model. See Pose2Sim for example: 
-    https://github.com/perfanalytics/pose2sim
-   
-    INPUTS:
-    - a video
-    - a dictionary obtained from a configuration file (.toml extension)
-    - a skeleton model
-    
-    OUTPUTS:
-    - one csv file of joint coordinates per detected person
-    - optionally json directory, image directory, video
-    - a logs.txt file 
-
-'''    
-
-
-## INIT
-import os
-import logging
-from pathlib import Path
-from sys import platform
-import json
-import subprocess
-import itertools as it
-import numpy as np
-import pandas as pd
-import cv2
-import matplotlib.pyplot as plt
-import matplotlib as mpl
-from Sports2D.Sports2D import base_params
-from Sports2D.Utilities import Blazepose_runsave, filter, common
-from Sports2D.Utilities.skeletons import *
-
-
-## AUTHORSHIP INFORMATION
-__author__ = "David Pagnon"
-__copyright__ = "Copyright 2023, Sports2D"
-__credits__ = ["David Pagnon"]
-__license__ = "BSD 3-Clause License"
-__version__ = "0.3.0"
-__maintainer__ = "David Pagnon"
-__email__ = "contact@david-pagnon.com"
-__status__ = "Development"
-
-
-# FUNCTIONS
-def display_figures_fun(df_list):
-    '''
-    Displays filtered and unfiltered data for comparison
-    /!\ Crashes on the third window...
-
-    INPUTS:
-    - df_list: list of dataframes of 3N columns, only 3i and 3i+1 are displayed
-
-    OUTPUT:
-    - matplotlib window with tabbed figures for each keypoint
-    '''
-    
-    mpl.use('qt5agg')
-    mpl.rc('figure', max_open_warning=0)
-
-    keypoints_names = df_list[0].columns.get_level_values(2)[1::3]
-    
-    pw = common.plotWindow()
-    for id, keypoint in enumerate(keypoints_names):
-        f = plt.figure()
-        
-        axX = plt.subplot(211)
-        [plt.plot(df.iloc[:,0], df.iloc[:,id*3+1], label=['unfiltered' if i==0 else 'filtered' if i==1 else ''][0]) for i,df in enumerate(df_list)]
[…]
