+++ Sports2D/__init__.py
@@ -3,6 +3,5 @@
 
 import sys
 
-
-__version__ = '0.1.0'
+__version__ = "0.4.0"
 VERSION = __version__

[…]
