### 2024-08-28 46d7449
All good (but colab), still needs to write Readme.mdand paper.md
