+++ Sports2D/process.py
@@ -7,38 +7,46 @@
     ## Compute pose and angles from video or webcam input       ##
     ##############################################################
     
-    Read video or webcam input
-    Detect joint centers with RTMPose
-    Attribute them to the same person across frames
-    Compute joint and segment angles
-    Optionally interpolate missing data, filter them, and display figures
-    Save image and video results, save pose as trc files, save angles as csv files
-    
-    Interpolates sequences of missing data if they are less than N frames long.
-    Optionally filters results with Butterworth, gaussian, median, or loess filter.
-    Optionally displays figures.
+    Detect 2D joint centers from a video or a webcam with RTMLib.
+    Compute selected joint and segment angles. 
+    Optionally save processed image files and video file.
+    Optionally save processed poses as a TRC file, and angles as a MOT file (OpenSim compatible).
 
-    If BlazePose is used, only one person can be detected.
-    No interpolation nor filtering options available. Not plotting available.
-
+    This scripts:
+    - loads skeleton information
+    - reads stream from a video or a webcam
+    - sets up the RTMLib pose tracker from RTMlib with specified parameters
+    - detects poses within the selected time range
+    - tracks people so that their IDs are consistent across frames
+    - retrieves the keypoints with high enough confidence, and only keeps the persons with enough high-confidence keypoints
+    - computes joint and segment angles, and flips those on the left/right side them if the respective foot is pointing to the left
+    - draws bounding boxes around each person with their IDs
+    - draws joint and segment angles on the body, and writes the values either near the joint/segment, or on the upper-left of the image with a progress bar
+    - draws the skeleton and the keypoints, with a green to red color scale to account for their confidence
+    - optionally show processed images, saves them, or saves them as a video
+    - interpolates missing pose and angle sequences if gaps are not too large
+    - filters them with the selected filter and parameters
+    - optionally plots pose and angle data before and after processing for comparison
+    - optionally saves poses for each person as a trc file, and angles as a mot file
+        
     /!\ Warning /!\
     - The pose detection is only as good as the pose estimation algorithm, i.e., it is not perfect.
-    - It will lead to reliable results only if the persons move in the 2D plane (sagittal plane).
+    - It will lead to reliable results only if the persons move in the 2D plane (sagittal or frontal plane).
     - The persons need to be filmed as perpendicularly as possible from their side.
     If you need research-grade markerless joint kinematics, consider using several cameras,
     and constraining angles to a biomechanically accurate model. See Pose2Sim for example: 
     https://github.com/perfanalytics/pose2sim
-   
+        
     INPUTS:
-    - a video
+    - a video or a webcam
     - a dictionary obtained from a configuration file (.toml extension)
     - a skeleton model
     
     OUTPUTS:
-    - one csv file of joint coordinates per detected person
-    - optionally json directory, image directory, video
+    - one trc file of joint coordinates per detected person
+    - one mot file of joint angles per detected person
+    - image files, video
[…]
