### 2024-08-28 3b514a3
entry points. Still needs to integrate comments
