+++ Sports2D/Sports2D.py
@@ -100,12 +100,80 @@
 
 
 ## INIT
+
+import argparse
 import toml
 from datetime import datetime
-import cv2
-import numpy as np
 from pathlib import Path
 import logging, logging.handlers
+import cv2
+import numpy as np
+
+from Sports2D import Sports2D
+
+
+## CONSTANTS
+DEFAULT_CONFIG =   {'project': {'video_input': 'Demo.mp4',
+                                'time_range': [],
+                                'video_dir': '',
+                                'webcam_id': 0,
+                                'input_size': [1280, 720]
+                                },
+                    'process': {'multiperson': True,
+                                'show_realtime_results': True,
+                                'save_vid': True,
+                                'save_img': True,
+                                'save_pose': True,
+                                'save_angles': True,
+                                'result_dir': ''
+                                },
+                    'pose':     { 'pose_model': 'body_with_feet',
+                                'mode': 'balanced',
+                                'det_frequency': 1,
+                                'tracking_mode': 'sports2d',
+                                'keypoint_likelihood_threshold': 0.3,
+                                'average_likelihood_threshold': 0.5,
+                                'keypoint_number_threshold': 0.3
+                                },
+                    'angles':   {'display_angle_values_on': ['body', 'list'],
+                                'joint_angles': [   'Right ankle',
+                                                    'Left ankle',
+                                                    'Right knee',
+                                                    'Left knee',
+                                                    'Right hip',
+                                                    'Left hip',
+                                                    'Right shoulder',
+                                                    'Left shoulder',
+                                                    'Right elbow',
+                                                    'Left elbow'],
+                                'segment_angles': [ 'Right foot',
+                                                    'Left foot',
+                                                    'Right shank',
+                                                    'Left shank',
+                                                    'Right thigh',
+                                                    'Left thigh',
+                                                    'Pelvis',
+                                                    'Trunk',
+                                                    'Shoulders',
+                                                    'Head',
+                                                    'Right arm',
+                                                    'Left arm',
+                                                    'Right forearm',
+                                                    'Left forearm'],
+                                'flip_left_right': True
+                                },
+                    'post-processing': {'interpolate': True,
+                                        'interp_gap_smaller_than': 10,
+                                        'fill_large_gaps_with': 'last_value',
+                                        'filter': True,
+                                        'show_plots': True,
+                                        'filter_type': 'butterworth',
+                                        'butterworth': {'order': 4, 'cut_off_frequency': 3},
+                                        'gaussian': {'sigma_kernel': 1},
+                                        'loess': {'nb_values_used': 5},
+                                        'median': {'kernel_size': 3}
+                                        }
+                    }
 
 
 ## AUTHORSHIP INFORMATION
@@ -181,6 +249,44 @@ def base_params(config_dict):
     return video_dir, video_files, frame_rates, time_ranges, result_dir
 
 
+def get_leaf_keys(config, prefix=''):
+    '''
+    Flatten configuration to map leaf keys to their full path
+    '''
+
+    leaf_keys = {}
[…]
