+++ README.md
@@ -97,7 +97,7 @@ If you need research-grade markerless joint kinematics, consider using several c
 
 ### Demonstration: Detect pose and compute 2D angles
 
-If you did the conda install, type `conda activate Sports2D` in the Anaconda prompt. Otherwise, skip to the next line.\
+If you installed via conda, type `conda activate Sports2D` in the Anaconda prompt. Otherwise, skip to the next line.\
 Open a terminal, enter `pip show sports2d`, check sports2d package location. \
 Copy this path and go to the Demo folder by typing `cd <path>\Sports2D\Demo`. \
 Type `ipython`, and test the following code:
