+++ .github/workflows/continuous-integration.yml
@@ -55,4 +55,4 @@ jobs:
       - name: Test with pytest
         run: |
           cd Sports2D/Utilities
-          pytest -v test_with_blazepose.py
+          pytest -v tests.py
