+++ setup.cfg
@@ -47,6 +47,9 @@ install_requires =
     imageio_ffmpeg
 packages = find:
 
+[options.entry_points]
+console_scripts = ['sports2d = Sports2D.Sports2D:main']
+
 [options.package_data]
 * = Demo/**/*, Demo/**/**/*
 
