+++ Sports2D/Sports2D.py
@@ -7,22 +7,34 @@
     ## SPORTS2D                                                 ##
     ##############################################################
     
-    This repository provides a workflow to compute 2D markerless
-    joint and segment angles from videos. 
-    These angles can be plotted and processed with any 
-    spreadsheet software or programming language.
-    
-    This is a headless version, but apps will be released 
-    for Windows, Linux, MacOS, as well as Android and iOS.
-    Mobile versions will only support exploratory joint detection 
-    from BlazePose, hence less accurately and less tunable.
-    
-    If you need to detect several persons and want more accurate results, 
-    you can install and use OpenPose: 
-    https://github.com/CMU-Perceptual-Computing-Lab/openpose
-    
+    Use sports2d to compute your athlete's pose, joint, and segment angles
+
     -----
-    Sports2D installation:
+    Help
+    -----
+    See https://github.com/davidpagnon/Sports2D
+    Or run: sports2d --help
+    Or check: https://github.com/davidpagnon/Sports2D/blob/main/Sports2D/Demo/Config_demo.toml
+
+    -----
+    Usage
+    -----
+    - Run on Demo video with default parameters: 
+        sports2d
+    - Run on custom video with default parameters:
+        sports2d --video_input path_to_video.mp4
+    - Run on multiple videos with default parameters:
+        sports2d --video_input path_to_video1.mp4 path_to_video2.mp4
+    - Run on webcam with default parameters: 
+        sports2d --video_input webcam
+    - Run with custom parameters (all non specified are set to default): 
+        sports2d --show_plots False --time_range 0 2.1 --result_dir path_to_result_dir
+        sports2d --multiperson false --mode lightweight --det_frequency 50
+    - Run with a toml configuration file: 
+        sports2d --config path_to_config.toml
+   
+    -----
+    Installation
     -----
     Optional: 
     - Install Miniconda
@@ -49,58 +61,61 @@
     https://github.com/perfanalytics/pose2sim
     
     -----
-    Pose detection:
+    How it works
     -----
-    Detect joint centers from a video with OpenPose or BlazePose.
-    Save a 2D csv position file per person, and optionally json files, image files, and video files.
-    
-    If OpenPose is used, multiple persons can be consistently detected across frames.
-    Interpolates sequences of missing data if they are less than N frames long.
-    Optionally filters results with Butterworth, gaussian, median, or loess filter.
-    Optionally displays figures.
+    Detects 2D joint centers from a video or a webcam with RTMLib.
+    Computes selected joint and segment angles. 
+    Optionally saves processed image files and video file.
+    Optionally saves processed poses as a TRC file, and angles as a MOT file (OpenSim compatible).
 
-    If BlazePose is used, only one person can be detected.
-    No interpolation nor filtering options available. Not plotting available.
-    
-    -----
-    Angle computation:
-    -----
-    Compute joint and segment angles from csv position files.
-    Automatically adjust angles when person switches to face the other way.
-    Save a 2D csv angle file per person.
-    Optionally filters results with Butterworth, gaussian, median, or loess filter.
-    Optionally displays figures.
-    Optionally saves images and video with overlaid angles.
+    Further details. Sports2D:
+    - loads skeleton information
+    - reads stream from a video or a webcam
+    - sets up the RTMLib pose tracker from RTMlib with specified parameters
+    - detects poses within the selected time range
+    - tracks people so that their IDs are consistent across frames
+    - retrieves the keypoints with high enough confidence, and only keeps the persons with enough high-confidence keypoints
[…]
