### 2024-08-30 10f727c
remove quotes from entry_points
