+++ setup.cfg
@@ -48,7 +48,7 @@ install_requires =
 packages = find:
 
 [options.entry_points]
-console_scripts = ['sports2d = Sports2D.Sports2D:main']
+console_scripts = [sports2d = Sports2D.Sports2D:main]
 
 [options.package_data]
 * = Demo/**/*, Demo/**/**/*
