### 2024-08-30 e318b0c
removed brackets around entroy points
