+++ Sports2D/process.py
@@ -74,38 +74,38 @@ from Sports2D.Utilities.skeletons import *
 ## CONSTANTS
 angle_dict = { 
     # joint angles
-    'Right ankle': [['RKnee', 'RAnkle', 'RBigToe', 'RHeel'], 'dorsiflexion', 90, 1],
-    'Left ankle': [['LKnee', 'LAnkle', 'LBigToe', 'LHeel'], 'dorsiflexion', 90, 1],
-    'Right knee': [['RAnkle', 'RKnee', 'RHip'], 'flexion', -180, 1],
-    'Left knee': [['LAnkle', 'LKnee', 'LHip'], 'flexion', -180, 1],
-    'Right hip': [['RKnee', 'RHip', 'Hip', 'Neck'], 'flexion', 0, -1],
-    'Left hip': [['LKnee', 'LHip', 'Hip', 'Neck'], 'flexion', 0, -1],
+    'right ankle': [['RKnee', 'RAnkle', 'RBigToe', 'RHeel'], 'dorsiflexion', 90, 1],
+    'left ankle': [['LKnee', 'LAnkle', 'LBigToe', 'LHeel'], 'dorsiflexion', 90, 1],
+    'right knee': [['RAnkle', 'RKnee', 'RHip'], 'flexion', -180, 1],
+    'left knee': [['LAnkle', 'LKnee', 'LHip'], 'flexion', -180, 1],
+    'right hip': [['RKnee', 'RHip', 'Hip', 'Neck'], 'flexion', 0, -1],
+    'left hip': [['LKnee', 'LHip', 'Hip', 'Neck'], 'flexion', 0, -1],
     # 'Lumbar': [['Neck', 'Hip', 'RHip', 'LHip'], 'flexion', -180, -1],
     # 'Neck': [['Head', 'Neck', 'RShoulder', 'LShoulder'], 'flexion', -180, -1],
-    'Right shoulder': [['RElbow', 'RShoulder', 'Hip', 'Neck'], 'flexion', 0, -1],
-    'Left shoulder': [['LElbow', 'LShoulder', 'Hip', 'Neck'], 'flexion', 0, -1],
-    'Right elbow': [['RWrist', 'RElbow', 'RShoulder'], 'flexion', 180, -1],
-    'Left elbow': [['LWrist', 'LElbow', 'LShoulder'], 'flexion', 180, -1],
-    'Right wrist': [['RElbow', 'RWrist', 'RIndex'], 'flexion', -180, 1],
-    'Left wrist': [['LElbow', 'LIndex', 'LWrist'], 'flexion', -180, 1],
+    'right shoulder': [['RElbow', 'RShoulder', 'Hip', 'Neck'], 'flexion', 0, -1],
+    'left shoulder': [['LElbow', 'LShoulder', 'Hip', 'Neck'], 'flexion', 0, -1],
+    'right elbow': [['RWrist', 'RElbow', 'RShoulder'], 'flexion', 180, -1],
+    'left elbow': [['LWrist', 'LElbow', 'LShoulder'], 'flexion', 180, -1],
+    'right wrist': [['RElbow', 'RWrist', 'RIndex'], 'flexion', -180, 1],
+    'left wrist': [['LElbow', 'LIndex', 'LWrist'], 'flexion', -180, 1],
 
     # segment angles
-    'Right foot': [['RBigToe', 'RHeel'], 'horizontal', 0, -1],
-    'Left foot': [['LBigToe', 'LHeel'], 'horizontal', 0, -1],
-    'Right shank': [['RAnkle', 'RKnee'], 'horizontal', 0, -1],
-    'Left shank': [['LAnkle', 'LKnee'], 'horizontal', 0, -1],
[…]
