+++ Sports2D/Utilities/tests.py
@@ -35,7 +35,7 @@ def test_workflow():
     '''
 
     # From python
-    config_dict = toml.load('../Config_demo.toml')
+    config_dict = toml.load('../Demo/Config_demo.toml')
     config_dict.get("process").update({"show_realtime_results":False})
     config_dict.get("post-processing").update({"show_plots":False})
     from Sports2D import Sports2D; Sports2D.process(config_dict)
