+++ setup.cfg
@@ -29,7 +29,7 @@ project_urls =
     
 [options]
 zip_safe = False
-python_requires =< 3.12
+python_requires <= 3.12
 install_requires =
     ipython
     toml
