### 2024-08-31 312fb9f
python version >=3.8 and <3.12 for MacOS + debug attempt
