+++ Sports2D/Sports2D.py
@@ -263,13 +263,14 @@ def base_params(config_dict):
         # video_files
         video_files = config_dict.get('project').get('video_input')
         if isinstance(video_files, str):
-            video_files = [video_dir / Path(video_files)]
+            video_files = [Path(video_files)]
         else: 
-            video_files = [video_dir / Path(v) for v in video_files]
+            video_files = [Path(v) for v in video_files]
 
         # frame_rates
         frame_rates = []
         for video_file in video_files:
+            print(str(video_dir / video_file))
             video = cv2.VideoCapture(str(video_dir / video_file))
             frame_rate = video.get(cv2.CAP_PROP_FPS)
             try:
