+++ .github/workflows/continuous-integration.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://docs.github.com/en/actions/automating-builds-and-tests/building-and-testing-python
 
-name: Build on Win-MacOS-Ubuntu with Python 3.8-3.10
+name: Build on Win-MacOS-Ubuntu with Python 3.8-3.11
 
 on:
   push:
@@ -23,7 +23,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest, windows-latest]
-        python-version: ["3.7", "3.8", "3.9", "3.10", "3.11", "3.12"]
+        python-version: ["3.8", "3.9", "3.10", "3.11"]
         include:
           - os: ubuntu-latest
             path: ~/.cache/pip
@@ -55,4 +55,4 @@ jobs:
       - name: Test with pytest
         run: |
           cd Sports2D/Utilities
-          pytest -v tests.py
+          pytest -v -s tests.py
