+++ Sports2D/Utilities/tests.py
@@ -36,7 +36,7 @@ def test_workflow():
 
     # From python
     config_dict = toml.load('../Demo/Config_demo.toml')
-    config_dict.get("process").update({"video_dir":'../Demo'})
+    config_dict.get("project").update({"video_dir":'../Demo'})
     config_dict.get("process").update({"show_realtime_results":False})
     config_dict.get("post-processing").update({"show_plots":False})
     
