### 2024-08-31 ed6c50b
troubleshooting ubuntu issue
