### 2024-08-31 655d952
fps not always retrieved successfully on Linux -> set to 30 if needed
