### 2024-08-31 #6849 @davidpagnon
The new version has been released!
- Used RTMPose rather than OpenPose or BlazePose (faster, more accurate, more flexible)
- Single process for both pose and angle estimation
- Webcam analysis available
- Body angle visualization
- Used entry points by default for clarity and simplicity
- Saved as OpenSim files: TRC for pose, MOT for angles
- Clearer documentation
- Better demo video
- Improved tracking IDs across frames
- Added some segment angles: Pelvis, Shoulders, Head

I am now working on editing the paper and fixing the Colab version.
