### 2024-09-01 7379b1b
making sure it throws an error when CLI fails
