+++ Sports2D/Utilities/tests.py
@@ -48,5 +48,5 @@ def test_workflow():
     subprocess.run(demo_cmd, check=True, capture_output=True, text=True)
     
     # From command line (CLI) with config file
-    demo_config_cmd = ["sports2d", "-c", "../Config_demo.toml", "--show_realtime_results", "False", "--show_plots", "False"]
+    demo_config_cmd = ["sports2d", "--config", "../Config_demo.toml", "--show_realtime_results", "False", "--show_plots", "False"]
     subprocess.run(demo_config_cmd, check=True, capture_output=True, text=True)
