### 2024-09-01 821b28d
fixed test with cli and config_path
