+++ Sports2D/Utilities/tests.py
@@ -51,5 +51,10 @@ def test_workflow():
     
     # From command line (CLI) with config file
     cli_config_path = Path(__file__).resolve().parent.parent / 'Demo' / 'Config_demo.toml'
+    cli_video_dir = Path(__file__).resolve().parent.parent / 'Demo'
+    config_dict = toml.load(cli_config_path)
+    config_dict.get("project").update({"video_dir": cli_video_dir})
+    with open(cli_config_path, 'w') as f: toml.dump(config_dict, f)
+    
     demo_config_cmd = ["sports2d", "--config", cli_config_path, "--show_realtime_results", "False", "--show_plots", "False"]
     subprocess.run(demo_config_cmd, check=True, capture_output=True, text=True)
