### 2024-09-01 9a6967b
Updated JOSS and Readme
