+++ Sports2D/Utilities/tests.py
@@ -45,14 +45,17 @@ def test_workflow():
     from Sports2D import Sports2D
     Sports2D.process(config_dict)
 
+
     # From command line (CLI)
     demo_cmd = ["sports2d", "--show_realtime_results", "False", "--show_plots", "False"]
     subprocess.run(demo_cmd, check=True, capture_output=True, text=True)
+  
     
     # From command line (CLI) with config file
     cli_config_path = Path(__file__).resolve().parent.parent / 'Demo' / 'Config_demo.toml'
     cli_video_file = Path(__file__).resolve().parent.parent / 'Demo' / 'demo.mp4'
     config_dict = toml.load(cli_config_path)
+    print(config_dict)
     config_dict.get("project").update({"video_input": cli_video_file})
     with open(cli_config_path, 'w') as f: toml.dump(config_dict, f)
 
