+++ Sports2D/Utilities/tests.py
@@ -53,10 +53,12 @@ def test_workflow():
     
     # From command line (CLI) with config file
     cli_config_path = Path(__file__).resolve().parent.parent / 'Demo' / 'Config_demo.toml'
-    cli_video_file = Path(__file__).resolve().parent.parent / 'Demo' / 'demo.mp4'
+    cli_video_file = 'demo.mp4'
+    cli_video_dir = Path(__file__).resolve().parent.parent / 'Demo'
     config_dict = toml.load(cli_config_path)
     print('cli_video_file: ', cli_video_file)
     config_dict.get("project").update({"video_input": cli_video_file})
+    config_dict.get("project").update({"video_dir": cli_video_dir})
     with open(cli_config_path, 'w') as f: toml.dump(config_dict, f)
     print('config_dict: ', config_dict)
 
