### 2024-09-02 fe9e87c
Small JOSS edits
