### 2024-09-02 1a6ebf3
is it just a path->str issue?
