### 2024-09-02 5afee05
fixes for webcam, plot angles, and added fontSize argument
