### 2024-09-02 b4f117e
yet another attempt to solve something that works but not on github actions
