+++ Sports2D/Utilities/tests.py
@@ -54,8 +54,10 @@ def test_workflow():
     # From command line (CLI) with config file
     cli_config_path = Path(__file__).resolve().parent.parent / 'Demo' / 'Config_demo.toml'
     config_dict = toml.load(cli_config_path)
-    cli_video_file = Path(__file__).resolve().parent.parent / 'Demo' / 'demo.mp4'
+    cli_video_file = 'demo.mp4'
+    cli_video_dir = Path(__file__).resolve().parent.parent / 'Demo'
     config_dict.get("project").update({"video_input": cli_video_file})
+    config_dict.get("project").update({"video_dir": cli_video_dir})
     with open(cli_config_path, 'w') as f: toml.dump(config_dict, f)
 
     demo_config_cmd = ["sports2d", "--config", str(cli_config_path), "--show_realtime_results", "False", "--show_plots", "False"]
