### 2024-09-02 396e433
Various fixes (Ubuntu, webcam, plot angles, fontsize)
