+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = sports2d
-version = 0.4.0
+version = 0.4.1
 author = David Pagnon
 author_email = contact@david-pagnon.com
 description = Detect pose and compute 2D joint angles from a video.
