### 2024-09-03 8e72205
case insensitive angles
