+++ Sports2D/Sports2D.py
@@ -192,8 +192,8 @@ CONFIG_HELP =   {'config': "Path to a toml configuration file",
                 'input_size': "width, height. 1280, 720 if not specified. Lower resolution will be faster but less precise",
                 'multiperson': "Multiperson involves tracking: will be faster if set to false. true if not specified",
                 'show_realtime_results': "show results in real-time. true if not specified",
-                'save_vid': "save video with overlaid angles. true if not specified",
-                'save_img': "save images with overlaid angles. true if not specified",
+                'save_vid': "save processed video. true if not specified",
+                'save_img': "save processed images. true if not specified",
                 'save_pose': "save pose as trc files. true if not specified",
                 'save_angles': "save angles as mot files. true if not specified",
                 'result_dir': "Current directory if not specified",
