### 2024-09-03 2677541
minor edit in sports2d -h
