+++ Sports2D/Sports2D.py
@@ -207,8 +207,8 @@ CONFIG_HELP =   {'config': "Path to a toml configuration file",
                 'keypoint_number_threshold': "Detected persons are not retained if number of detected keypoints is below this threshold. 0.3 if not specified, i.e., i.e., 30 percent",
                 'display_angle_values_on': "body, list, or body list. body list if not specified",
                 'fontSize': "Font size for angle values. 0.3 if not specified",
-                'joint_angles': "Right ankle, Left ankle, Right knee, Left knee, Right hip, Left hip, Right shoulder, Left shoulder, Right elbow, Left elbow if not specified",
-                'segment_angles': "Right foot, Left foot, Right shank, Left shank, Right thigh, Left thigh, Pelvis, Trunk, Shoulders, Head, Right arm, Left arm, Right forearm, Left forearm if not specified",
+                'joint_angles': '"Right ankle" "Left ankle" "Right knee" "Left knee" "Right hip" "Left hip" "Right shoulder" "Left shoulder" "Right elbow" "Left elbow" if not specified',
+                'segment_angles': '"Right foot" "Left foot" "Right shank" "Left shank" "Right thigh" "Left thigh" "Pelvis" "Trunk" "Shoulders" "Head" "Right arm" "Left arm" "Right forearm" "Left forearm" if not specified',
                 'flip_left_right': "true or false. true to get consistent angles with people facing both left and right sides. Set it to false if you want timeseries to be continuous even when the participent switches their stance. true if not specified",
                 'interpolate': "Interpolate missing data. true if not specified",
                 'interp_gap_smaller_than': "Interpolate sequences of missing data if they are less than N frames long. 10 if not specified",
