### 2024-09-03 08de0e0
Adjust fontSize dynamically
