### 2024-09-03 db4d29d
adding comments and changing scale factors
