### 2024-09-03 c930314
scale factor = 1.5
