### 2024-09-04 83bdd5f
fix when video_dir not specified
