### 2024-09-05 5079041
Checking for fontSize is str
