### 2024-09-05 4d60edd
comment for isinstance
