### 2024-09-05 4c52de8
shorter key available (-i instead of --video_input)
