+++ Sports2D/Sports2D.py
@@ -175,7 +175,7 @@ DEFAULT_CONFIG =   {'project': {'video_input': ['demo.mp4'],
                                         'interp_gap_smaller_than': 10,
                                         'fill_large_gaps_with': 'last_value',
                                         'filter': True,
-                                        'show_plots': True,
+                                        'show_graphs': True,
                                         'filter_type': 'butterworth',
                                         'butterworth': {'order': 4, 'cut_off_frequency': 3},
                                         'gaussian': {'sigma_kernel': 1},
@@ -184,43 +184,43 @@ DEFAULT_CONFIG =   {'project': {'video_input': ['demo.mp4'],
                                         }
                     }
 
-CONFIG_HELP =   {'config': "Path to a toml configuration file",
-                'video_input': "webcam, or video_path.mp4, or video1_path.avi video2_path.mp4 ... Beware that images won't be saved if paths contain non ASCII characters",
-                'time_range': "start_time, end_time. In seconds. Whole video if not specified",
-                'video_dir': "Current directory if not specified",
-                'webcam_id': "webcam ID. 0 if not specified",
-                'input_size': "width, height. 1280, 720 if not specified. Lower resolution will be faster but less precise",
-                'multiperson': "Multiperson involves tracking: will be faster if set to false. true if not specified",
-                'show_realtime_results': "show results in real-time. true if not specified",
-                'save_vid': "save processed video. true if not specified",
-                'save_img': "save processed images. true if not specified",
-                'save_pose': "save pose as trc files. true if not specified",
-                'save_angles': "save angles as mot files. true if not specified",
-                'result_dir': "Current directory if not specified",
-                'pose_model': "Only body_with_feet is available for now. body_with_feet if not specified",
-                'mode': "light, balanced, or performance. balanced if not specified",
-                'det_frequency': "Run person detection only every N frames, and inbetween track previously detected bounding boxes. keypoint detection is still run on all frames.\n\
-                                 Equal to or greater than 1, can be as high as you want in simple uncrowded cases. Much faster, but might be less accurate. 1 if not specified: detection runs on all frames",
-                'tracking_mode': "sports2d or rtmlib. sports2d is generally much more accurate and comparable in speed. sports2d if not specified",
-                'keypoint_likelihood_threshold': "Detected keypoints are not retained if likelihood is below this threshold. 0.3 if not specified",
-                'average_likelihood_threshold': "Detected persons are not retained if average keypoint likelihood is below this threshold. 0.5 if not specified",
-                'keypoint_number_threshold': "Detected persons are not retained if number of detected keypoints is below this threshold. 0.3 if not specified, i.e., i.e., 30 percent",
-                'display_angle_values_on': "body, list, or body list. body list if not specified",
-                'fontSize': "Font size for angle values. 0.3 if not specified",
-                'joint_angles': '"Right ankle" "Left ankle" "Right knee" "Left knee" "Right hip" "Left hip" "Right shoulder" "Left shoulder" "Right elbow" "Left elbow" if not specified',
-                'segment_angles': '"Right foot" "Left foot" "Right shank" "Left shank" "Right thigh" "Left thigh" "Pelvis" "Trunk" "Shoulders" "Head" "Right arm" "Left arm" "Right forearm" "Left forearm" if not specified',
[…]
