### 2024-09-05 78d462c
changed show_plots to show_graphs (conflicting short arguments)
