+++ Sports2D/Sports2D.py
@@ -28,7 +28,7 @@
     - Run on webcam with default parameters: 
         sports2d --video_input webcam
     - Run with custom parameters (all non specified are set to default): 
-        sports2d --show_plots False --time_range 0 2.1 --result_dir path_to_result_dir
+        sports2d --show_graphs False --time_range 0 2.1 --result_dir path_to_result_dir
         sports2d --multiperson false --mode lightweight --det_frequency 50
     - Run with a toml configuration file: 
         sports2d --config path_to_config.toml
@@ -407,7 +407,7 @@ def main():
     - Run on webcam with default parameters: 
         sports2d --video_input webcam
     - Run with custom parameters (all non specified are set to default): 
-        sports2d --show_plots False --time_range 0 2.1 --result_dir path_to_result_dir
+        sports2d --show_graphs False --time_range 0 2.1 --result_dir path_to_result_dir
         sports2d --multiperson false --mode lightweight --det_frequency 50
     - Run with a toml configuration file: 
         sports2d --config path_to_config.toml
