### 2024-09-06 adb5429
fixed time_range = [] in DEFAULT_CONFIG
