### 2024-09-06 8f0212e
More information on GPU
