+++ Sports2D/process.py
@@ -132,7 +132,7 @@ def setup_webcam(webcam_id, save_vid, vid_output_path, input_size):
 
     INPUTS:
     - webcam_id: int. The ID of the webcam to capture from
-    - input_size: tuple. The size of the input frame (width, height)
+    - input_size: tuple. The size of the input frame (width, height)\
 
     OUTPUTS:
     - cap: cv2.VideoCapture. The webcam capture object
@@ -478,6 +478,27 @@ def sort_people_rtmlib(pose_tracker, keypoints, scores):
 
     return sorted_keypoints, sorted_scores
 
+def dynamic_fontSize(width, height, base_fontSize=0.3, base_dimension=1768):
+    '''
+    Dynamically adjust font size according to the max dimension (width or height).
+    '''
+    # Max dimension
+    max_dimension = max(width, height)
+
+    # Calculate scale
+    scale = max_dimension / base_dimension
+    
+    # Adjust font size with a fixed scale factor of 1.2 for better readability
+    adjusted_fontSize = base_fontSize * scale * 1.2
+    
+    # Ensure font size is within the allowed range but, not used parameters for boundaries
+    return max(min(adjusted_fontSize, 1), 0.2)
+
+def dynamic_thickness(fontSize, scale_factor=5):
+    '''
+    Dynamically adjust the thickness of the lines according to the font size.
+    '''
+    return max(int(fontSize * scale_factor), 1) # Minimum thickness is 1.
 
 def draw_dotted_line(img, start, direction, length, color=(0, 255, 0), gap=7, dot_length=3, thickness=1):
     '''
@@ -500,7 +521,7 @@ def draw_dotted_line(img, start, direction, length, color=(0, 255, 0), gap=7, do
     for i in range(0, length, gap):
         line_start = start + direction * i
         line_end = line_start + direction * dot_length
-        cv2.line(img, tuple(line_start.astype(int)), tuple(line_end.astype(int)), color, thickness)
+        cv2.line(img, tuple(line_start.astype(int)), tuple(line_end.astype(int)), color, thickness+1 if thickness<2 else thickness)
 
 
 def draw_bounding_box(img, X, Y, colors=[(255, 0, 0), (0, 255, 0), (0, 0, 255)], fontSize=0.3, thickness=1):
@@ -568,7 +589,7 @@ def draw_skel(img, X, Y, model, colors=[(255, 0, 0), (0, 255, 0), (0, 0, 255)], 
         c = next(color_cycle)
         if not np.isnan(x).all():
             [cv2.line(img,
-                (int(x[n[0]]), int(y[n[0]])), (int(x[n[1]]), int(y[n[1]])), c, thickness)
+                (int(x[n[0]]), int(y[n[0]])), (int(x[n[1]]), int(y[n[1]])), c, thickness+1 if thickness<2 else thickness)
                 for n in node_pairs
                 if not (np.isnan(x[n[0]]) or np.isnan(y[n[0]]) or np.isnan(x[n[1]]) or np.isnan(y[n[1]]))]
 
@@ -627,17 +648,17 @@ def draw_angles(img, valid_X, valid_Y, valid_angles, valid_X_flipped, keypoints_
     - img: image with angles
     '''
 
-    color_cycle = it.cycle(colors)
+    color_cycle = it.cycle(colors) 
     for person_id, (X,Y,angles, X_flipped) in enumerate(zip(valid_X, valid_Y, valid_angles, valid_X_flipped)):
         c = next(color_cycle)
         if not np.isnan(X).all():
             # person label
             if 'list' in display_angle_values_on:
                 person_label_position = (int(10 + fontSize*150/0.3*person_id), int(fontSize*50))
[…]
