### 2024-09-08 86139b9
Thickness of bbox
