### 2024-09-13 #6849 @nicos1993
@davidpagnon @Kevin-Mattheus-Moerman Thanks for the changes and updates made, it looks great. I tried the demo again post updates and its cool! I have 1 suggestion that you can wish to ignore and 1 comment.
_Suggestion comment_
For the plots that get generated automatically it would be nice if they are labeled with the matching subject ID.
_Comment_
Does the Google Colab functionality work yet? It is mentioned in your revised article, but on your repository it is still 'unticked' in the How To Contribute section. I do not wish to tick all the remaining checklist boxes if it is claimed to work as per the article but does not actually work yet (I do not think it works yet). If that is the case then I would remove that text from the article, and I would be satisfied to accept it.

Thanks

Nicos
