### 2024-09-16 68e5b86
switched plots and terminal images
