### 2024-09-16 f0087f8
updated paper (remove mention to colab version)
