+++ .github/workflows/joss_pdf.yml
@@ -20,7 +20,7 @@ jobs:
           # This should be the path to the paper within your repo.
           paper-path: Content/paper.md
       - name: Upload
-        uses: actions/upload-artifact@v1
+        uses: actions/upload-artifact@v4
         with:
           name: paper
           # This is the output path where Pandoc will write the compiled
