### 2024-09-17 #6849 @davidpagnon
Thanks Nicos!
- Good suggestion for the plot titles, it is now done
- Yes, it should not be hard to make the Colab version work but unfortunately, I won't have time to do it over the next few days, so I'll just remove any mention of it from the paper.
