### 2024-09-23 f0407f3
Update paper.md
