### 2024-09-23 #6849 @davidpagnon
## Paper:
[x] UK -> United Kingdom
[x] Matched Zenodo and JOSS info

## Latest pypi release:
[0.4.6](https://pypi.org/project/sports2d/)

## Zenodo archive DOI:
[10.5281/zenodo.7903962](https://zenodo.org/doi/10.5281/zenodo.7903962)
Added a second author to the Zenodo archive
