### 2024-09-24 #6849 @Kevin-Mattheus-Moerman
@davidpagnon can you update the archive listed license to match your project license?

Currently it lists this default:
![image](https://github.com/user-attachments/assets/10def7bd-909b-40b8-bbde-02c537602848)
