### 2024-09-24 #6849 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
