### 2024-09-24 0daf6df
typo + slight edit on last paragraph of statement of needs
