### 2024-09-24 #6849 @davidpagnon
I noticed a typo (two stars ** that had to be removed) and slightly rephrased something in the last paragraph of the statement of need. It probably does not matter if this is too late.
