### 2024-09-24 #6849 @Kevin-Mattheus-Moerman
@davidpagnon I now speak to you as an AEiC for JOSS to help process the work for acceptance. I will now process some final checks:

- [x] Software license tag listed here matches a tagged release
