### 2024-09-24 fccc280
Update CITATION.cff
