### 2024-09-24 9b7c50f
Update citation upon JOSS acceptance
