### 2024-09-26 #6849 @Kevin-Mattheus-Moerman
@davidpagnon congratulations on this JOSS publication.

I'd like to say thanks to reviewers: @tuliofalmeida, @nicos1993 !!!!
